Split Vulkan renderer into modular translation units
