+++ meson.build
@@ -220,6 +220,10 @@ elif renderer_opt == 'renderer_vk'
   refresh_src = [
     'src/renderer_vk/main.cpp',
     'src/renderer_vk/renderer.cpp',
+    'src/renderer_vk/resources.cpp',
+    'src/renderer_vk/device.cpp',
+    'src/renderer_vk/frame.cpp',
+    'src/renderer_vk/ui.cpp',
     'src/renderer_vk/vk_draw2d.cpp',
   ]
 endif
