+++ src/renderer_vk/renderer.cpp
@@ -1,4931 +1,10 @@
 #include "renderer.h"
 
-#include "vk_draw2d.h"
-
-#include "renderer/common.h"
-#include "renderer/images.h"
-#include "renderer/kfont.h"
-
-#include <algorithm>
-#include <array>
-#include <atomic>
-#include <cmath>
-#include <cstdint>
-#include <cstdlib>
-#include <cstring>
-#include <limits>
-#include <string>
-#include <unordered_set>
-
-#include "common/cmodel.h"
-#include "common/files.h"
-#include "client/client.h"
-
-extern uint32_t d_8to24table[256];
-
 namespace refresh::vk {
 
-namespace {
-    uint8_t colorBitsForFormat(VkFormat format) {
-        switch (format) {
-        case VK_FORMAT_B8G8R8A8_UNORM:
-        case VK_FORMAT_B8G8R8A8_SRGB:
-        case VK_FORMAT_R8G8B8A8_UNORM:
-        case VK_FORMAT_R8G8B8A8_SRGB:
-            return 24;
-        case VK_FORMAT_B8G8R8_UNORM:
-        case VK_FORMAT_R8G8B8_UNORM:
-            return 24;
-        case VK_FORMAT_B5G6R5_UNORM_PACK16:
-        case VK_FORMAT_R5G6B5_UNORM_PACK16:
-            return 16;
-        case VK_FORMAT_A2B10G10R10_UNORM_PACK32:
-        case VK_FORMAT_A2R10G10B10_UNORM_PACK32:
-            return 30;
-        default:
-            return 24;
-        }
-    }
-
-    constexpr int kDefaultCharWidth = 8;
-    constexpr int kDefaultCharHeight = 8;
-    constexpr float kInverseLightIntensity = 1.0f / 255.0f;
-    constexpr float kFontCellSize = 1.0f / 16.0f;
-    constexpr float kTileDivisor = 1.0f / 64.0f;
-    constexpr float kShadowOffset = 1.0f;
-    constexpr int kUIDropShadow = 1 << 4;
-    constexpr int kUIAltColor = 1 << 5;
-    constexpr int kUIXorColor = 1 << 7;
-
-    struct VideoGeometry {
-        int width = SCREEN_WIDTH;
-        int height = SCREEN_HEIGHT;
-        vidFlags_t flags = {};
-    };
-
-    int countPrintable(std::string_view value, size_t maxChars) {
-        size_t count = 0;
-        for (char ch : value) {
-            if (ch == '\0') {
-                break;
-            }
-            if (maxChars && count >= maxChars) {
-                break;
-            }
-            ++count;
-        }
-        return static_cast<int>(count);
-    }
-
-    std::array<float, 3> toArray(const vec3_t value) {
-        return { value[0], value[1], value[2] };
-    }
-
-    constexpr float kParticleSize = 1.0f + M_SQRT1_2f;
-    constexpr float kParticleScale = 1.0f / (2.0f * kParticleSize);
-    constexpr float kParticleDistanceBias = 20.0f;
-    constexpr float kParticleDistanceScale = 0.004f;
-    constexpr int kBeamCylinderSides = 12;
-
-    std::array<float, 4> toColorArray(color_t color, float alphaScale = 1.0f) {
-        constexpr float kInv255 = 1.0f / 255.0f;
-        float alpha = static_cast<float>(color.a) * kInv255 * alphaScale;
-        alpha = std::clamp(alpha, 0.0f, 1.0f);
-        return {
-            static_cast<float>(color.r) * kInv255,
-            static_cast<float>(color.g) * kInv255,
-            static_cast<float>(color.b) * kInv255,
-            alpha,
-        };
-    }
-
-    void arrayToVec3(const std::array<float, 3> &src, vec3_t dst) {
-        dst[0] = src[0];
-        dst[1] = src[1];
-        dst[2] = src[2];
-    }
-
-    VulkanRenderer::ModelRecord::AliasFrameMetadata makeAliasFrameMetadata(const maliasframe_t &frame) {
-        VulkanRenderer::ModelRecord::AliasFrameMetadata metadata{};
-        metadata.boundsMin = toArray(frame.bounds[0]);
-        metadata.boundsMax = toArray(frame.bounds[1]);
-        metadata.scale = toArray(frame.scale);
-        metadata.translate = toArray(frame.translate);
-        metadata.radius = frame.radius;
-        return metadata;
-    }
-
-    std::array<std::array<float, 2>, 4> makeQuad(float x, float y, float w, float h) {
-        return {{{
-            {x, y},
-            {x + w, y},
-            {x + w, y + h},
-            {x, y + h},
-        }}};
-    }
-
-    std::array<std::array<float, 2>, 4> makeUV(float u0, float v0, float u1, float v1) {
-        return {{{
-            {u0, v0},
-            {u1, v0},
-            {u1, v1},
-            {u0, v1},
-        }}};
-    }
-
-    const std::array<std::array<float, 2>, 4> kFullUVs = makeUV(0.0f, 0.0f, 1.0f, 1.0f);
-    std::array<std::array<float, 2>, 4> makeQuadPositions(float x0, float y0, float x1, float y1) {
-        return {{{ { x0, y0 }, { x1, y0 }, { x1, y1 }, { x0, y1 } }}};
-    }
-
-    cvar_t *vk_fog = nullptr;
-    cvar_t *vk_bloom = nullptr;
-    cvar_t *vk_polyblend = nullptr;
-    cvar_t *vk_waterwarp = nullptr;
-    cvar_t *vk_dynamic = nullptr;
-    cvar_t *vk_perPixelLighting = nullptr;
-
-    bool legacyToggleValue(const char *name, bool defaultValue) {
-        if (!name || !*name) {
-            return defaultValue;
-        }
-        if (cvar_t *legacy = Cvar_FindVar(name)) {
-            return legacy->integer > 0;
-        }
-        return defaultValue;
-    }
-
-    bool resolveToggle(cvar_t *primary, const char *legacyName, bool defaultValue) {
-        bool fallback = legacyToggleValue(legacyName, defaultValue);
-        if (!primary) {
-            return fallback;
-        }
-        int value = primary->integer;
-        if (value < 0) {
-            return fallback;
-        }
-        return value > 0;
-    }
-
-    std::string describeFogBits(refresh::vk::VulkanRenderer::FogBits bits) {
-        using FogBits = refresh::vk::VulkanRenderer::FogBits;
-        if (bits == FogBits::FogNone) {
-            return {};
-        }
-
-        std::string description{"("};
-        bool first = true;
-        auto append = [&](const char *label) {
-            if (!first) {
-                description.append("|");
-            }
-            description.append(label);
-            first = false;
-        };
-
-        if ((bits & FogBits::FogGlobal) != FogBits::FogNone) {
-            append("global");
-        }
-        if ((bits & FogBits::FogHeight) != FogBits::FogNone) {
-            append("height");
-        }
-        if ((bits & FogBits::FogSky) != FogBits::FogNone) {
-            append("sky");
-        }
-
-        if (first) {
-            return {};
-        }
-        description.push_back(')');
-        return description;
-    }
-
-    VideoGeometry queryVideoGeometry() {
-        VideoGeometry geometry{};
-
-        vrect_t rect{};
-        if (VID_GetGeometry(&rect)) {
-            geometry.width = std::max(1, rect.width);
-            geometry.height = std::max(1, rect.height);
-        }
-
-        vrect_t fullscreen{};
-        int freq = 0;
-        int depth = 0;
-        if (VID_GetFullscreen(&fullscreen, &freq, &depth)) {
-            geometry.flags = static_cast<vidFlags_t>(geometry.flags | QVF_FULLSCREEN);
-        }
-
-        return geometry;
-    }
-
-    void applyVideoGeometry(const VideoGeometry &geometry) {
-        r_config.width = geometry.width;
-        r_config.height = geometry.height;
-        r_config.flags = geometry.flags;
-    }
-}
-
-model_t *MOD_Find(const char *name);
-
-void VulkanRenderer::RenderQueues::clear() {
-    beams.clear();
-    flares.clear();
-    bmodels.clear();
-    opaque.clear();
-    alphaBack.clear();
-    alphaFront.clear();
-}
-
-void VulkanRenderer::FramePrimitiveBuffers::clear() {
-    beams.clear();
-    particles.clear();
-    flares.clear();
-    debugLines.clear();
-}
-
-void VulkanRenderer::EffectVertexStreams::clear() {
-    beamVertices.clear();
-    beamIndices.clear();
-    particleVertices.clear();
-    flareVertices.clear();
-    flareIndices.clear();
-    debugLinesDepth.clear();
-    debugLinesNoDepth.clear();
-}
-
-void VulkanRenderer::FrameStats::reset() {
-    drawCalls = 0;
-    pipelinesBound = 0;
-    beams = 0;
-    particles = 0;
-    flares = 0;
-    debugLines = 0;
-}
-
 VulkanRenderer::VulkanRenderer()
-    : handleCounter_{1} {
-}
+    : handleCounter_{1} {}
 
 VulkanRenderer::~VulkanRenderer() = default;
-
-qhandle_t VulkanRenderer::nextHandle() {
-    return handleCounter_.fetch_add(1, std::memory_order_relaxed);
-}
-
-qhandle_t VulkanRenderer::registerResource(NameLookup &lookup, std::string_view name) {
-    if (name.empty()) {
-        return 0;
-    }
-
-    std::string key{name};
-
-    if (auto it = lookup.find(key); it != lookup.end()) {
-        return it->second;
-    }
-
-    qhandle_t handle = nextHandle();
-    lookup.emplace(std::move(key), handle);
-    return handle;
-}
-
-bool VulkanRenderer::createTextureDescriptorSetLayout() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    if (textureDescriptorSetLayout_ != VK_NULL_HANDLE) {
-        return true;
-    }
-
-    VkDescriptorSetLayoutBinding binding{};
-    binding.binding = 0;
-    binding.descriptorCount = 1;
-    binding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    binding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-
-    VkDescriptorSetLayoutCreateInfo info{};
-    info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    info.bindingCount = 1;
-    info.pBindings = &binding;
-
-    VkResult result = vkCreateDescriptorSetLayout(device_, &info, nullptr, &textureDescriptorSetLayout_);
-    if (result != VK_SUCCESS || textureDescriptorSetLayout_ == VK_NULL_HANDLE) {
-        textureDescriptorSetLayout_ = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to create texture descriptor set layout (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    return true;
-}
-
-void VulkanRenderer::destroyTextureDescriptorSetLayout() {
-    if (device_ == VK_NULL_HANDLE) {
-        textureDescriptorSetLayout_ = VK_NULL_HANDLE;
-        return;
-    }
-
-    if (textureDescriptorSetLayout_ != VK_NULL_HANDLE) {
-        vkDestroyDescriptorSetLayout(device_, textureDescriptorSetLayout_, nullptr);
-        textureDescriptorSetLayout_ = VK_NULL_HANDLE;
-    }
-}
-
-void VulkanRenderer::destroyImageRecord(ImageRecord &record) {
-    if (device_ == VK_NULL_HANDLE) {
-        record.image = VK_NULL_HANDLE;
-        record.memory = VK_NULL_HANDLE;
-        record.view = VK_NULL_HANDLE;
-        record.sampler = VK_NULL_HANDLE;
-        record.descriptorSet = VK_NULL_HANDLE;
-        record.layout = VK_IMAGE_LAYOUT_UNDEFINED;
-        return;
-    }
-
-    if (record.descriptorSet != VK_NULL_HANDLE && descriptorPool_ != VK_NULL_HANDLE) {
-        vkFreeDescriptorSets(device_, descriptorPool_, 1, &record.descriptorSet);
-        record.descriptorSet = VK_NULL_HANDLE;
-    }
-
-    if (record.sampler != VK_NULL_HANDLE) {
-        vkDestroySampler(device_, record.sampler, nullptr);
-        record.sampler = VK_NULL_HANDLE;
-    }
-
-    if (record.view != VK_NULL_HANDLE) {
-        vkDestroyImageView(device_, record.view, nullptr);
-        record.view = VK_NULL_HANDLE;
-    }
-
-    if (record.image != VK_NULL_HANDLE) {
-        vkDestroyImage(device_, record.image, nullptr);
-        record.image = VK_NULL_HANDLE;
-    }
-
-    if (record.memory != VK_NULL_HANDLE) {
-        vkFreeMemory(device_, record.memory, nullptr);
-        record.memory = VK_NULL_HANDLE;
-    }
-
-    record.layout = VK_IMAGE_LAYOUT_UNDEFINED;
-    record.extent = { 0u, 0u, 1u };
-    record.format = VK_FORMAT_UNDEFINED;
-}
-
-void VulkanRenderer::destroyAllImageResources() {
-    for (auto &entry : images_) {
-        destroyImageRecord(entry.second);
-    }
-}
-
-bool VulkanRenderer::allocateTextureDescriptor(ImageRecord &record) {
-    if (device_ == VK_NULL_HANDLE || descriptorPool_ == VK_NULL_HANDLE || textureDescriptorSetLayout_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    VkDescriptorSetAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-    allocInfo.descriptorPool = descriptorPool_;
-    allocInfo.descriptorSetCount = 1;
-    allocInfo.pSetLayouts = &textureDescriptorSetLayout_;
-
-    VkResult result = vkAllocateDescriptorSets(device_, &allocInfo, &record.descriptorSet);
-    if (result != VK_SUCCESS || record.descriptorSet == VK_NULL_HANDLE) {
-        record.descriptorSet = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to allocate texture descriptor set (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    VkDescriptorImageInfo imageInfo{};
-    imageInfo.imageLayout = record.layout;
-    imageInfo.imageView = record.view;
-    imageInfo.sampler = record.sampler;
-
-    VkWriteDescriptorSet write{};
-    write.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-    write.dstSet = record.descriptorSet;
-    write.dstBinding = 0;
-    write.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    write.descriptorCount = 1;
-    write.pImageInfo = &imageInfo;
-
-    vkUpdateDescriptorSets(device_, 1, &write, 0, nullptr);
-    return true;
-}
-
-uint32_t VulkanRenderer::findMemoryType(uint32_t typeBits, VkMemoryPropertyFlags properties) const {
-    for (uint32_t i = 0; i < memoryProperties_.memoryTypeCount; ++i) {
-        if ((typeBits & (1u << i)) && (memoryProperties_.memoryTypes[i].propertyFlags & properties) == properties) {
-            return i;
-        }
-    }
-    return UINT32_MAX;
-}
-
-bool VulkanRenderer::createBuffer(VkDeviceSize size, VkBufferUsageFlags usage, VkMemoryPropertyFlags properties,
-                                  VkBuffer &buffer, VkDeviceMemory &memory) {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    VkBufferCreateInfo info{};
-    info.sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO;
-    info.size = size;
-    info.usage = usage;
-    info.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    VkResult result = vkCreateBuffer(device_, &info, nullptr, &buffer);
-    if (result != VK_SUCCESS || buffer == VK_NULL_HANDLE) {
-        buffer = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to create buffer (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    VkMemoryRequirements requirements{};
-    vkGetBufferMemoryRequirements(device_, buffer, &requirements);
-
-    uint32_t memoryType = findMemoryType(requirements.memoryTypeBits, properties);
-    if (memoryType == UINT32_MAX) {
-        vkDestroyBuffer(device_, buffer, nullptr);
-        buffer = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to find suitable memory type for buffer.\n");
-        return false;
-    }
-
-    VkMemoryAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    allocInfo.allocationSize = requirements.size;
-    allocInfo.memoryTypeIndex = memoryType;
-
-    result = vkAllocateMemory(device_, &allocInfo, nullptr, &memory);
-    if (result != VK_SUCCESS || memory == VK_NULL_HANDLE) {
-        vkDestroyBuffer(device_, buffer, nullptr);
-        buffer = VK_NULL_HANDLE;
-        memory = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to allocate buffer memory (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    result = vkBindBufferMemory(device_, buffer, memory, 0);
-    if (result != VK_SUCCESS) {
-        vkDestroyBuffer(device_, buffer, nullptr);
-        vkFreeMemory(device_, memory, nullptr);
-        buffer = VK_NULL_HANDLE;
-        memory = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to bind buffer memory (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    return true;
-}
-
-void VulkanRenderer::destroyBuffer(VkBuffer buffer, VkDeviceMemory memory) {
-    if (device_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    if (buffer != VK_NULL_HANDLE) {
-        vkDestroyBuffer(device_, buffer, nullptr);
-    }
-    if (memory != VK_NULL_HANDLE) {
-        vkFreeMemory(device_, memory, nullptr);
-    }
-}
-
-VkCommandBuffer VulkanRenderer::beginSingleTimeCommands() {
-    if (device_ == VK_NULL_HANDLE || commandPool_ == VK_NULL_HANDLE) {
-        return VK_NULL_HANDLE;
-    }
-
-    VkCommandBufferAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    allocInfo.commandPool = commandPool_;
-    allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    allocInfo.commandBufferCount = 1;
-
-    VkCommandBuffer commandBuffer = VK_NULL_HANDLE;
-    VkResult result = vkAllocateCommandBuffers(device_, &allocInfo, &commandBuffer);
-    if (result != VK_SUCCESS || commandBuffer == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to allocate command buffer for upload (VkResult %d).\n", static_cast<int>(result));
-        return VK_NULL_HANDLE;
-    }
-
-    VkCommandBufferBeginInfo beginInfo{};
-    beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-    beginInfo.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-
-    result = vkBeginCommandBuffer(commandBuffer, &beginInfo);
-    if (result != VK_SUCCESS) {
-        vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-        Com_Printf("refresh-vk: failed to begin upload command buffer (VkResult %d).\n", static_cast<int>(result));
-        return VK_NULL_HANDLE;
-    }
-
-    return commandBuffer;
-}
-
-void VulkanRenderer::endSingleTimeCommands(VkCommandBuffer commandBuffer) {
-    if (device_ == VK_NULL_HANDLE || commandBuffer == VK_NULL_HANDLE) {
-        return;
-    }
-
-    vkEndCommandBuffer(commandBuffer);
-
-    VkSubmitInfo submitInfo{};
-    submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-    submitInfo.commandBufferCount = 1;
-    submitInfo.pCommandBuffers = &commandBuffer;
-
-    VkResult result = vkQueueSubmit(graphicsQueue_, 1, &submitInfo, VK_NULL_HANDLE);
-    if (result != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to submit upload command buffer (VkResult %d).\n", static_cast<int>(result));
-    }
-
-    vkQueueWaitIdle(graphicsQueue_);
-    vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-}
-
-void VulkanRenderer::transitionImageLayout(VkCommandBuffer commandBuffer, VkImage image,
-                                           VkImageLayout oldLayout, VkImageLayout newLayout) {
-    if (commandBuffer == VK_NULL_HANDLE || image == VK_NULL_HANDLE) {
-        return;
-    }
-
-    VkImageMemoryBarrier barrier{};
-    barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-    barrier.oldLayout = oldLayout;
-    barrier.newLayout = newLayout;
-    barrier.srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.image = image;
-    barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    barrier.subresourceRange.baseMipLevel = 0;
-    barrier.subresourceRange.levelCount = 1;
-    barrier.subresourceRange.baseArrayLayer = 0;
-    barrier.subresourceRange.layerCount = 1;
-
-    VkPipelineStageFlags srcStage = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-    VkPipelineStageFlags dstStage = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-
-    if (oldLayout == VK_IMAGE_LAYOUT_UNDEFINED && newLayout == VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL) {
-        barrier.srcAccessMask = 0;
-        barrier.dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-        srcStage = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-        dstStage = VK_PIPELINE_STAGE_TRANSFER_BIT;
-    } else if (oldLayout == VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL && newLayout == VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL) {
-        barrier.srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-        barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-        srcStage = VK_PIPELINE_STAGE_TRANSFER_BIT;
-        dstStage = VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT;
-    }
-
-    vkCmdPipelineBarrier(commandBuffer,
-                         srcStage, dstStage,
-                         0,
-                         0, nullptr,
-                         0, nullptr,
-                         1, &barrier);
-}
-
-void VulkanRenderer::copyBufferToImage(VkCommandBuffer commandBuffer, VkBuffer buffer, VkImage image,
-                                       uint32_t width, uint32_t height) {
-    if (commandBuffer == VK_NULL_HANDLE || buffer == VK_NULL_HANDLE || image == VK_NULL_HANDLE) {
-        return;
-    }
-
-    VkBufferImageCopy region{};
-    region.bufferOffset = 0;
-    region.bufferRowLength = 0;
-    region.bufferImageHeight = 0;
-    region.imageSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    region.imageSubresource.mipLevel = 0;
-    region.imageSubresource.baseArrayLayer = 0;
-    region.imageSubresource.layerCount = 1;
-    region.imageOffset = { 0, 0, 0 };
-    region.imageExtent = { width, height, 1 };
-
-    vkCmdCopyBufferToImage(commandBuffer, buffer, image, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, 1, &region);
-}
-
-bool VulkanRenderer::uploadImagePixels(ImageRecord &record, const uint8_t *pixels, size_t size,
-                                       uint32_t width, uint32_t height, VkFormat format) {
-    if (!pixels || size == 0 || width == 0 || height == 0) {
-        return false;
-    }
-
-    VkDeviceSize dataSize = static_cast<VkDeviceSize>(size);
-    VkBuffer stagingBuffer = VK_NULL_HANDLE;
-    VkDeviceMemory stagingMemory = VK_NULL_HANDLE;
-
-    if (!createBuffer(dataSize,
-                      VK_BUFFER_USAGE_TRANSFER_SRC_BIT,
-                      VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT,
-                      stagingBuffer,
-                      stagingMemory)) {
-        return false;
-    }
-
-    void *mapped = nullptr;
-    VkResult result = vkMapMemory(device_, stagingMemory, 0, dataSize, 0, &mapped);
-    if (result != VK_SUCCESS || mapped == nullptr) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        Com_Printf("refresh-vk: failed to map staging buffer memory (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    std::memcpy(mapped, pixels, static_cast<size_t>(dataSize));
-    vkUnmapMemory(device_, stagingMemory);
-
-    VkImageCreateInfo imageInfo{};
-    imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-    imageInfo.imageType = VK_IMAGE_TYPE_2D;
-    imageInfo.extent.width = width;
-    imageInfo.extent.height = height;
-    imageInfo.extent.depth = 1;
-    imageInfo.mipLevels = 1;
-    imageInfo.arrayLayers = 1;
-    imageInfo.format = format;
-    imageInfo.tiling = VK_IMAGE_TILING_OPTIMAL;
-    imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    imageInfo.usage = VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT;
-    imageInfo.samples = VK_SAMPLE_COUNT_1_BIT;
-    imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    result = vkCreateImage(device_, &imageInfo, nullptr, &record.image);
-    if (result != VK_SUCCESS || record.image == VK_NULL_HANDLE) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        record.image = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to create image (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    VkMemoryRequirements requirements{};
-    vkGetImageMemoryRequirements(device_, record.image, &requirements);
-
-    uint32_t memoryType = findMemoryType(requirements.memoryTypeBits, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-    if (memoryType == UINT32_MAX) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        vkDestroyImage(device_, record.image, nullptr);
-        record.image = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to find suitable memory type for image.\n");
-        return false;
-    }
-
-    VkMemoryAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    allocInfo.allocationSize = requirements.size;
-    allocInfo.memoryTypeIndex = memoryType;
-
-    result = vkAllocateMemory(device_, &allocInfo, nullptr, &record.memory);
-    if (result != VK_SUCCESS || record.memory == VK_NULL_HANDLE) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        vkDestroyImage(device_, record.image, nullptr);
-        record.image = VK_NULL_HANDLE;
-        record.memory = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to allocate image memory (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    result = vkBindImageMemory(device_, record.image, record.memory, 0);
-    if (result != VK_SUCCESS) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        vkDestroyImage(device_, record.image, nullptr);
-        vkFreeMemory(device_, record.memory, nullptr);
-        record.image = VK_NULL_HANDLE;
-        record.memory = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to bind image memory (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    VkCommandBuffer commandBuffer = beginSingleTimeCommands();
-    if (commandBuffer == VK_NULL_HANDLE) {
-        destroyBuffer(stagingBuffer, stagingMemory);
-        return false;
-    }
-
-    transitionImageLayout(commandBuffer, record.image, VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL);
-    copyBufferToImage(commandBuffer, stagingBuffer, record.image, width, height);
-    transitionImageLayout(commandBuffer, record.image, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
-
-    endSingleTimeCommands(commandBuffer);
-
-    destroyBuffer(stagingBuffer, stagingMemory);
-
-    VkImageViewCreateInfo viewInfo{};
-    viewInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-    viewInfo.image = record.image;
-    viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-    viewInfo.format = format;
-    viewInfo.components = {
-        VK_COMPONENT_SWIZZLE_IDENTITY,
-        VK_COMPONENT_SWIZZLE_IDENTITY,
-        VK_COMPONENT_SWIZZLE_IDENTITY,
-        VK_COMPONENT_SWIZZLE_IDENTITY,
-    };
-    viewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    viewInfo.subresourceRange.baseMipLevel = 0;
-    viewInfo.subresourceRange.levelCount = 1;
-    viewInfo.subresourceRange.baseArrayLayer = 0;
-    viewInfo.subresourceRange.layerCount = 1;
-
-    result = vkCreateImageView(device_, &viewInfo, nullptr, &record.view);
-    if (result != VK_SUCCESS || record.view == VK_NULL_HANDLE) {
-        vkDestroyImage(device_, record.image, nullptr);
-        vkFreeMemory(device_, record.memory, nullptr);
-        record.image = VK_NULL_HANDLE;
-        record.memory = VK_NULL_HANDLE;
-        record.view = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to create image view (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    VkSamplerCreateInfo samplerInfo{};
-    samplerInfo.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-    samplerInfo.magFilter = VK_FILTER_LINEAR;
-    samplerInfo.minFilter = VK_FILTER_LINEAR;
-    samplerInfo.addressModeU = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-    samplerInfo.addressModeV = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-    samplerInfo.addressModeW = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-    samplerInfo.anisotropyEnable = VK_FALSE;
-    samplerInfo.maxAnisotropy = 1.0f;
-    samplerInfo.borderColor = VK_BORDER_COLOR_INT_OPAQUE_BLACK;
-    samplerInfo.unnormalizedCoordinates = VK_FALSE;
-    samplerInfo.compareEnable = VK_FALSE;
-    samplerInfo.mipmapMode = VK_SAMPLER_MIPMAP_MODE_LINEAR;
-    samplerInfo.mipLodBias = 0.0f;
-    samplerInfo.minLod = 0.0f;
-    samplerInfo.maxLod = 0.0f;
-
-    result = vkCreateSampler(device_, &samplerInfo, nullptr, &record.sampler);
-    if (result != VK_SUCCESS || record.sampler == VK_NULL_HANDLE) {
-        vkDestroyImageView(device_, record.view, nullptr);
-        vkDestroyImage(device_, record.image, nullptr);
-        vkFreeMemory(device_, record.memory, nullptr);
-        record.view = VK_NULL_HANDLE;
-        record.image = VK_NULL_HANDLE;
-        record.memory = VK_NULL_HANDLE;
-        record.sampler = VK_NULL_HANDLE;
-        Com_Printf("refresh-vk: failed to create sampler (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    record.layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    record.extent = { width, height, 1 };
-    record.format = format;
-    return true;
-}
-
-bool VulkanRenderer::ensureTextureResources(ImageRecord &record, const uint8_t *pixels, size_t size,
-                                            uint32_t width, uint32_t height, VkFormat format) {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    destroyImageRecord(record);
-
-    if (!uploadImagePixels(record, pixels, size, width, height, format)) {
-        destroyImageRecord(record);
-        return false;
-    }
-
-    if (!allocateTextureDescriptor(record)) {
-        destroyImageRecord(record);
-        return false;
-    }
-
-    return true;
-}
-
-void VulkanRenderer::resetTransientState() {
-    clipRect_.reset();
-    activeScissor_.reset();
-    scale_ = 1.0f;
-    autoScaleValue_ = 1;
-    clear2DBatches();
-    resetFrameState();
-    lastSubmittedFrame_.reset();
-}
-
-int VulkanRenderer::computeAutoScale() const {
-    const int baseHeight = SCREEN_HEIGHT;
-    const int baseWidth = SCREEN_WIDTH;
-
-    int width = std::max(1, r_config.width);
-    int height = std::max(1, r_config.height);
-
-    int scale = 1;
-    if (height >= width) {
-        scale = width / baseWidth;
-    } else {
-        scale = height / baseHeight;
-    }
-
-    if (scale < 1) {
-        scale = 1;
-    }
-
-    if (vid && vid->get_dpi_scale) {
-        int dpiScale = vid->get_dpi_scale();
-        if (dpiScale > scale) {
-            scale = dpiScale;
-        }
-    }
-
-    return scale;
-}
-
-void VulkanRenderer::updateUIScaling() {
-    autoScaleValue_ = std::max(1, computeAutoScale());
-
-    if (!std::isfinite(scale_) || scale_ <= 0.0f) {
-        scale_ = 1.0f;
-    }
-
-    // Re-apply the current scale to refresh dependent state such as clip rectangles.
-    setScale(scale_);
-}
-
-int VulkanRenderer::readSwapIntervalSetting() const {
-    int desired = 1;
-    if (gl_swapinterval) {
-        desired = gl_swapinterval->integer;
-    } else {
-        if (cvar_t *swap = Cvar_Find("gl_swapinterval")) {
-            desired = swap->integer;
-        }
-    }
-
-    if (desired < 0) {
-        desired = 0;
-    }
-
-    return desired;
-}
-
-bool VulkanRenderer::refreshSwapInterval(bool allowRecreate) {
-    int desired = readSwapIntervalSetting();
-    if (desired == swapInterval_) {
-        return false;
-    }
-
-    swapInterval_ = desired;
-
-    if (swapInterval_ > 0) {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags | QVF_VIDEOSYNC);
-    } else {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags & ~QVF_VIDEOSYNC);
-    }
-
-    if (allowRecreate && initialized_ && device_ != VK_NULL_HANDLE && swapchain_ != VK_NULL_HANDLE) {
-        rebuildSwapchain();
-    }
-
-    return true;
-}
-
-void VulkanRenderer::clearFrameTransientQueues() {
-    frameQueues_.clear();
-    framePrimitives_.clear();
-    effectStreams_.clear();
-}
-
-void VulkanRenderer::resetFrameStatistics() {
-    frameStats_.reset();
-}
-
-void VulkanRenderer::recordStage(std::string_view label) {
-    commandLog_.emplace_back(label);
-}
-
-void VulkanRenderer::recordDrawCall(const PipelineDesc &pipeline, std::string_view label, size_t count) {
-    std::string entry{label};
-    entry.append(" [");
-    entry.append(pipeline.debugName);
-    entry.push_back(']');
-    if (count > 0) {
-        entry.push_back(' ');
-        entry.push_back('x');
-        entry.append(std::to_string(count));
-    }
-    commandLog_.push_back(std::move(entry));
-    frameStats_.drawCalls += 1;
-}
-
-VulkanRenderer::PipelineDesc VulkanRenderer::makePipeline(const PipelineKey &key) const {
-    PipelineDesc desc{};
-    desc.key = key;
-
-    switch (key.kind) {
-    case PipelineKind::InlineBsp:
-        desc.debugName = "inline_bsp";
-        break;
-    case PipelineKind::Alias:
-        desc.debugName = "alias";
-        break;
-    case PipelineKind::Sprite:
-        desc.debugName = "sprite";
-        break;
-    case PipelineKind::Weapon:
-        desc.debugName = "weapon";
-        break;
-    case PipelineKind::Draw2D:
-        desc.debugName = "draw2d";
-        break;
-    case PipelineKind::BeamSimple:
-        desc.debugName = "beam.simple";
-        desc.blend = PipelineDesc::BlendMode::Alpha;
-        desc.depthWrite = false;
-        desc.textured = true;
-        break;
-    case PipelineKind::BeamCylindrical:
-        desc.debugName = "beam.cylinder";
-        desc.blend = PipelineDesc::BlendMode::Alpha;
-        desc.depthWrite = false;
-        desc.textured = false;
-        break;
-    case PipelineKind::ParticleAlpha:
-        desc.debugName = "particle.alpha";
-        desc.blend = PipelineDesc::BlendMode::Alpha;
-        desc.depthWrite = false;
-        desc.textured = true;
-        break;
-    case PipelineKind::ParticleAdditive:
-        desc.debugName = "particle.add";
-        desc.blend = PipelineDesc::BlendMode::Additive;
-        desc.depthWrite = false;
-        desc.textured = true;
-        break;
-    case PipelineKind::Flare:
-        desc.debugName = "flare";
-        desc.blend = PipelineDesc::BlendMode::Additive;
-        desc.depthTest = false;
-        desc.depthWrite = false;
-        desc.textured = true;
-        break;
-    case PipelineKind::DebugLineDepth:
-        desc.debugName = "debug.line.depth";
-        desc.topology = VK_PRIMITIVE_TOPOLOGY_LINE_LIST;
-        desc.blend = PipelineDesc::BlendMode::Alpha;
-        desc.depthWrite = false;
-        desc.textured = false;
-        break;
-    case PipelineKind::DebugLineNoDepth:
-        desc.debugName = "debug.line.nodepth";
-        desc.topology = VK_PRIMITIVE_TOPOLOGY_LINE_LIST;
-        desc.blend = PipelineDesc::BlendMode::Alpha;
-        desc.depthTest = false;
-        desc.depthWrite = false;
-        desc.textured = false;
-        break;
-    }
-
-    desc.usesFog = key.fogBits != FogNone;
-    desc.usesSkyFog = key.fogSkyBits != FogNone;
-    desc.usesDynamicLights = key.perPixelLighting && key.dynamicLights;
-
-    auto appendFogSuffix = [&](const char *label, FogBits bits) {
-        std::string fog = describeFogBits(bits);
-        if (!fog.empty()) {
-            desc.debugName.push_back('.');
-            desc.debugName.append(label);
-            desc.debugName.append(fog);
-        }
-    };
-
-    appendFogSuffix("fog", key.fogBits);
-    if (key.fogSkyBits != FogNone) {
-        appendFogSuffix("sky", key.fogSkyBits);
-    }
-    if (key.perPixelLighting) {
-        desc.debugName.append(".ppl");
-    }
-    if (key.dynamicLights) {
-        desc.debugName.append(".dl");
-    }
-
-    return desc;
-}
-
-VulkanRenderer::PipelineKey VulkanRenderer::buildPipelineKey(PipelineKind kind) const {
-    PipelineKey key{};
-    key.kind = kind;
-
-    switch (kind) {
-    case PipelineKind::InlineBsp:
-        key.fogBits = frameState_.fogBits;
-        key.fogSkyBits = frameState_.fogBitsSky;
-        key.perPixelLighting = frameState_.perPixelLighting;
-        key.dynamicLights = frameState_.dynamicLightsUploaded;
-        break;
-    case PipelineKind::Alias:
-    case PipelineKind::Sprite:
-    case PipelineKind::Weapon:
-        key.fogBits = frameState_.fogBits;
-        key.perPixelLighting = frameState_.perPixelLighting;
-        key.dynamicLights = frameState_.dynamicLightsUploaded;
-        break;
-    default:
-        break;
-    }
-
-    return key;
-}
-
-const VulkanRenderer::PipelineDesc &VulkanRenderer::ensurePipeline(const PipelineKey &key) {
-    if (auto it = pipelines_.find(key); it != pipelines_.end()) {
-        return it->second;
-    }
-
-    PipelineDesc desc = makePipeline(key);
-    auto [it, inserted] = pipelines_.emplace(key, std::move(desc));
-    if (inserted) {
-        frameStats_.pipelinesBound += 1;
-    }
-    return it->second;
-}
-
-const VulkanRenderer::ModelRecord *VulkanRenderer::findModelRecord(qhandle_t handle) const {
-    if (auto it = models_.find(handle); it != models_.end()) {
-        return &it->second;
-    }
-    return nullptr;
-}
-
-VulkanRenderer::ModelRecord *VulkanRenderer::findModelRecord(qhandle_t handle) {
-    if (auto it = models_.find(handle); it != models_.end()) {
-        return &it->second;
-    }
-    return nullptr;
-}
-
-const VulkanRenderer::ImageRecord *VulkanRenderer::findImageRecord(qhandle_t handle) const {
-    if (auto it = images_.find(handle); it != images_.end()) {
-        return &it->second;
-    }
-    return nullptr;
-}
-
-const VulkanRenderer::KFontRecord *VulkanRenderer::findKFontRecord(const kfont_t *font) const {
-    if (!font) {
-        return nullptr;
-    }
-
-    if (auto it = kfontCache_.find(font); it != kfontCache_.end()) {
-        return &it->second;
-    }
-    return nullptr;
-}
-
-std::string_view VulkanRenderer::classifyModelName(const ModelRecord *record) const {
-    if (!record || record->name.empty()) {
-        return {};
-    }
-    const std::string &name = record->name;
-    size_t dot = name.find_last_of('.');
-    if (dot == std::string::npos || dot + 1 >= name.size()) {
-        return {};
-    }
-    return std::string_view{name}.substr(dot);
-}
-
-uint32_t VulkanRenderer::findMemoryType(uint32_t typeFilter, VkMemoryPropertyFlags properties) const {
-    if (physicalDevice_ == VK_NULL_HANDLE) {
-        return UINT32_MAX;
-    }
-
-    VkPhysicalDeviceMemoryProperties memoryProperties{};
-    vkGetPhysicalDeviceMemoryProperties(physicalDevice_, &memoryProperties);
-
-    for (uint32_t i = 0; i < memoryProperties.memoryTypeCount; ++i) {
-        if ((typeFilter & (1u << i)) == 0) {
-            continue;
-        }
-        if ((memoryProperties.memoryTypes[i].propertyFlags & properties) == properties) {
-            return i;
-        }
-    }
-
-    return UINT32_MAX;
-}
-
-bool VulkanRenderer::createBuffer(ModelRecord::BufferAllocationInfo &buffer,
-                                  VkDeviceSize size,
-                                  VkBufferUsageFlags usage,
-                                  VkMemoryPropertyFlags properties) {
-    if (device_ == VK_NULL_HANDLE || size == 0) {
-        return false;
-    }
-
-    VkBufferCreateInfo bufferInfo{};
-    bufferInfo.sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO;
-    bufferInfo.size = size;
-    bufferInfo.usage = usage;
-    bufferInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    VkResult bufferResult = vkCreateBuffer(device_, &bufferInfo, nullptr, &buffer.buffer);
-    if (bufferResult != VK_SUCCESS || buffer.buffer == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create buffer (VkResult %d).\n", static_cast<int>(bufferResult));
-        buffer.buffer = VK_NULL_HANDLE;
-        return false;
-    }
-
-    VkMemoryRequirements requirements{};
-    vkGetBufferMemoryRequirements(device_, buffer.buffer, &requirements);
-
-    VkMemoryAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    allocInfo.allocationSize = requirements.size;
-    allocInfo.memoryTypeIndex = findMemoryType(requirements.memoryTypeBits, properties);
-
-    if (allocInfo.memoryTypeIndex == UINT32_MAX) {
-        Com_Printf("refresh-vk: unable to find compatible memory type for buffer.\n");
-        vkDestroyBuffer(device_, buffer.buffer, nullptr);
-        buffer.buffer = VK_NULL_HANDLE;
-        return false;
-    }
-
-    VkResult memoryResult = vkAllocateMemory(device_, &allocInfo, nullptr, &buffer.memory);
-    if (memoryResult != VK_SUCCESS || buffer.memory == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to allocate buffer memory (VkResult %d).\n", static_cast<int>(memoryResult));
-        vkDestroyBuffer(device_, buffer.buffer, nullptr);
-        buffer.buffer = VK_NULL_HANDLE;
-        buffer.memory = VK_NULL_HANDLE;
-        return false;
-    }
-
-    VkResult bindResult = vkBindBufferMemory(device_, buffer.buffer, buffer.memory, 0);
-    if (bindResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to bind buffer memory (VkResult %d).\n", static_cast<int>(bindResult));
-        vkDestroyBuffer(device_, buffer.buffer, nullptr);
-        vkFreeMemory(device_, buffer.memory, nullptr);
-        buffer.buffer = VK_NULL_HANDLE;
-        buffer.memory = VK_NULL_HANDLE;
-        return false;
-    }
-
-    buffer.offset = 0;
-    buffer.size = size;
-    return true;
-}
-
-void VulkanRenderer::destroyBuffer(ModelRecord::BufferAllocationInfo &buffer) {
-    if (device_ == VK_NULL_HANDLE) {
-        buffer = {};
-        return;
-    }
-
-    if (buffer.buffer != VK_NULL_HANDLE) {
-        vkDestroyBuffer(device_, buffer.buffer, nullptr);
-        buffer.buffer = VK_NULL_HANDLE;
-    }
-
-    if (buffer.memory != VK_NULL_HANDLE) {
-        vkFreeMemory(device_, buffer.memory, nullptr);
-        buffer.memory = VK_NULL_HANDLE;
-    }
-
-    buffer.offset = 0;
-    buffer.size = 0;
-}
-
-void VulkanRenderer::destroy2DBatch(Draw2DBatch &batch) {
-    if (device_ != VK_NULL_HANDLE && descriptorPool_ != VK_NULL_HANDLE && batch.descriptor.set != VK_NULL_HANDLE) {
-        vkFreeDescriptorSets(device_, descriptorPool_, 1, &batch.descriptor.set);
-    }
-    batch.descriptor.set = VK_NULL_HANDLE;
-    batch.descriptor.vertex = {};
-    batch.descriptor.index = {};
-
-    destroyBuffer(batch.vertex);
-    destroyBuffer(batch.index);
-
-    batch.vertexCount = 0;
-    batch.indexCount = 0;
-    batch.texture = 0;
-}
-
-void VulkanRenderer::clear2DBatches() {
-    for (auto &batch : frame2DBatches_) {
-        destroy2DBatch(batch);
-    }
-    frame2DBatches_.clear();
-}
-
-bool VulkanRenderer::copyBuffer(VkBuffer src, VkBuffer dst, VkDeviceSize size) {
-    if (device_ == VK_NULL_HANDLE || commandPool_ == VK_NULL_HANDLE || graphicsQueue_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    VkCommandBufferAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    allocInfo.commandPool = commandPool_;
-    allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    allocInfo.commandBufferCount = 1;
-
-    VkCommandBuffer commandBuffer = VK_NULL_HANDLE;
-    VkResult allocResult = vkAllocateCommandBuffers(device_, &allocInfo, &commandBuffer);
-    if (allocResult != VK_SUCCESS || commandBuffer == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to allocate command buffer for copy (VkResult %d).\n", static_cast<int>(allocResult));
-        return false;
-    }
-
-    VkCommandBufferBeginInfo beginInfo{};
-    beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-    beginInfo.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-
-    VkResult beginResult = vkBeginCommandBuffer(commandBuffer, &beginInfo);
-    if (beginResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to begin copy command buffer (VkResult %d).\n", static_cast<int>(beginResult));
-        vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-        return false;
-    }
-
-    VkBufferCopy copyRegion{};
-    copyRegion.srcOffset = 0;
-    copyRegion.dstOffset = 0;
-    copyRegion.size = size;
-    vkCmdCopyBuffer(commandBuffer, src, dst, 1, &copyRegion);
-
-    VkResult endResult = vkEndCommandBuffer(commandBuffer);
-    if (endResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to end copy command buffer (VkResult %d).\n", static_cast<int>(endResult));
-        vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-        return false;
-    }
-
-    VkSubmitInfo submitInfo{};
-    submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-    submitInfo.commandBufferCount = 1;
-    submitInfo.pCommandBuffers = &commandBuffer;
-
-    VkResult submitResult = vkQueueSubmit(graphicsQueue_, 1, &submitInfo, VK_NULL_HANDLE);
-    if (submitResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to submit buffer copy (VkResult %d).\n", static_cast<int>(submitResult));
-        vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-        return false;
-    }
-
-    VkResult waitResult = vkQueueWaitIdle(graphicsQueue_);
-    if (waitResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: queue wait failed after buffer copy (VkResult %d).\n", static_cast<int>(waitResult));
-        vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-        return false;
-    }
-
-    vkFreeCommandBuffers(device_, commandPool_, 1, &commandBuffer);
-    return true;
-}
-
-bool VulkanRenderer::createModelDescriptorResources() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    if (modelDescriptorSetLayout_ == VK_NULL_HANDLE) {
-        VkDescriptorSetLayoutBinding vertexBinding{};
-        vertexBinding.binding = 0;
-        vertexBinding.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-        vertexBinding.descriptorCount = 1;
-        vertexBinding.stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
-
-        VkDescriptorSetLayoutBinding indexBinding = vertexBinding;
-        indexBinding.binding = 1;
-
-        std::array<VkDescriptorSetLayoutBinding, 2> bindings{vertexBinding, indexBinding};
-
-        VkDescriptorSetLayoutCreateInfo layoutInfo{};
-        layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-        layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
-        layoutInfo.pBindings = bindings.data();
-
-        VkResult layoutResult = vkCreateDescriptorSetLayout(device_, &layoutInfo, nullptr, &modelDescriptorSetLayout_);
-        if (layoutResult != VK_SUCCESS || modelDescriptorSetLayout_ == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create model descriptor set layout (VkResult %d).\n", static_cast<int>(layoutResult));
-            modelDescriptorSetLayout_ = VK_NULL_HANDLE;
-            return false;
-        }
-    }
-
-    if (modelPipelineLayout_ == VK_NULL_HANDLE) {
-        VkPipelineLayoutCreateInfo pipelineInfo{};
-        pipelineInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-        pipelineInfo.setLayoutCount = 1;
-        pipelineInfo.pSetLayouts = &modelDescriptorSetLayout_;
-
-        VkResult pipelineResult = vkCreatePipelineLayout(device_, &pipelineInfo, nullptr, &modelPipelineLayout_);
-        if (pipelineResult != VK_SUCCESS || modelPipelineLayout_ == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create model pipeline layout (VkResult %d).\n", static_cast<int>(pipelineResult));
-            if (modelPipelineLayout_ != VK_NULL_HANDLE) {
-                vkDestroyPipelineLayout(device_, modelPipelineLayout_, nullptr);
-                modelPipelineLayout_ = VK_NULL_HANDLE;
-            }
-            if (modelDescriptorSetLayout_ != VK_NULL_HANDLE) {
-                vkDestroyDescriptorSetLayout(device_, modelDescriptorSetLayout_, nullptr);
-                modelDescriptorSetLayout_ = VK_NULL_HANDLE;
-            }
-            return false;
-        }
-    }
-
-    return true;
-}
-
-void VulkanRenderer::destroyModelDescriptorResources() {
-    if (device_ == VK_NULL_HANDLE) {
-        modelPipelineLayout_ = VK_NULL_HANDLE;
-        modelDescriptorSetLayout_ = VK_NULL_HANDLE;
-        return;
-    }
-
-    if (modelPipelineLayout_ != VK_NULL_HANDLE) {
-        vkDestroyPipelineLayout(device_, modelPipelineLayout_, nullptr);
-        modelPipelineLayout_ = VK_NULL_HANDLE;
-    }
-
-    if (modelDescriptorSetLayout_ != VK_NULL_HANDLE) {
-        vkDestroyDescriptorSetLayout(device_, modelDescriptorSetLayout_, nullptr);
-        modelDescriptorSetLayout_ = VK_NULL_HANDLE;
-    }
-}
-
-bool VulkanRenderer::uploadMeshGeometry(ModelRecord::MeshGeometry &geometry) {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    const bool hasVertexData = !geometry.vertexStaging.empty();
-    const bool hasIndexData = !geometry.indexStaging.empty();
-
-    if (!hasVertexData && !hasIndexData) {
-        geometry.uploaded = true;
-        return true;
-    }
-
-    ModelRecord::BufferAllocationInfo newVertex{};
-    ModelRecord::BufferAllocationInfo newIndex{};
-
-    if (hasVertexData) {
-        ModelRecord::BufferAllocationInfo staging{};
-        VkDeviceSize vertexSize = static_cast<VkDeviceSize>(geometry.vertexStaging.size());
-        if (!createBuffer(staging,
-                          vertexSize,
-                          VK_BUFFER_USAGE_TRANSFER_SRC_BIT,
-                          VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT)) {
-            return false;
-        }
-
-        void *mapped = nullptr;
-        VkResult mapResult = vkMapMemory(device_, staging.memory, staging.offset, staging.size, 0, &mapped);
-        if (mapResult != VK_SUCCESS || mapped == nullptr) {
-            Com_Printf("refresh-vk: failed to map vertex staging memory (VkResult %d).\n", static_cast<int>(mapResult));
-            destroyBuffer(staging);
-            return false;
-        }
-
-        std::memcpy(mapped, geometry.vertexStaging.data(), geometry.vertexStaging.size());
-        vkUnmapMemory(device_, staging.memory);
-
-        if (!createBuffer(newVertex,
-                          vertexSize,
-                          VK_BUFFER_USAGE_VERTEX_BUFFER_BIT | VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                          VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)) {
-            destroyBuffer(staging);
-            return false;
-        }
-
-        if (!copyBuffer(staging.buffer, newVertex.buffer, newVertex.size)) {
-            destroyBuffer(staging);
-            destroyBuffer(newVertex);
-            return false;
-        }
-
-        destroyBuffer(staging);
-    }
-
-    if (hasIndexData) {
-        ModelRecord::BufferAllocationInfo staging{};
-        VkDeviceSize indexSize = static_cast<VkDeviceSize>(geometry.indexStaging.size());
-        if (!createBuffer(staging,
-                          indexSize,
-                          VK_BUFFER_USAGE_TRANSFER_SRC_BIT,
-                          VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT)) {
-            if (hasVertexData) {
-                destroyBuffer(newVertex);
-            }
-            return false;
-        }
-
-        void *mapped = nullptr;
-        VkResult mapResult = vkMapMemory(device_, staging.memory, staging.offset, staging.size, 0, &mapped);
-        if (mapResult != VK_SUCCESS || mapped == nullptr) {
-            Com_Printf("refresh-vk: failed to map index staging memory (VkResult %d).\n", static_cast<int>(mapResult));
-            destroyBuffer(staging);
-            if (hasVertexData) {
-                destroyBuffer(newVertex);
-            }
-            return false;
-        }
-
-        std::memcpy(mapped, geometry.indexStaging.data(), geometry.indexStaging.size());
-        vkUnmapMemory(device_, staging.memory);
-
-        if (!createBuffer(newIndex,
-                          indexSize,
-                          VK_BUFFER_USAGE_INDEX_BUFFER_BIT | VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                          VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)) {
-            destroyBuffer(staging);
-            if (hasVertexData) {
-                destroyBuffer(newVertex);
-            }
-            return false;
-        }
-
-        if (!copyBuffer(staging.buffer, newIndex.buffer, newIndex.size)) {
-            destroyBuffer(staging);
-            destroyBuffer(newIndex);
-            if (hasVertexData) {
-                destroyBuffer(newVertex);
-            }
-            return false;
-        }
-
-        destroyBuffer(staging);
-    }
-
-    if (hasVertexData) {
-        destroyBuffer(geometry.vertex);
-        geometry.vertex = newVertex;
-        geometry.descriptor.vertex.buffer = geometry.vertex.buffer;
-        geometry.descriptor.vertex.offset = geometry.vertex.offset;
-        geometry.descriptor.vertex.range = geometry.vertex.size;
-    }
-
-    if (hasIndexData) {
-        destroyBuffer(geometry.index);
-        geometry.index = newIndex;
-        geometry.descriptor.index.buffer = geometry.index.buffer;
-        geometry.descriptor.index.offset = geometry.index.offset;
-        geometry.descriptor.index.range = geometry.index.size;
-    }
-
-    if ((geometry.vertex.buffer != VK_NULL_HANDLE || geometry.index.buffer != VK_NULL_HANDLE) &&
-        descriptorPool_ != VK_NULL_HANDLE) {
-        if (geometry.descriptor.set == VK_NULL_HANDLE) {
-            if (createModelDescriptorResources()) {
-                VkDescriptorSetAllocateInfo allocInfo{};
-                allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-                allocInfo.descriptorPool = descriptorPool_;
-                allocInfo.descriptorSetCount = 1;
-                allocInfo.pSetLayouts = &modelDescriptorSetLayout_;
-
-                VkResult allocResult = vkAllocateDescriptorSets(device_, &allocInfo, &geometry.descriptor.set);
-                if (allocResult != VK_SUCCESS) {
-                    Com_Printf("refresh-vk: failed to allocate model descriptor set (VkResult %d).\n", static_cast<int>(allocResult));
-                    geometry.descriptor.set = VK_NULL_HANDLE;
-                }
-            }
-        }
-
-        if (geometry.descriptor.set != VK_NULL_HANDLE) {
-            std::array<VkWriteDescriptorSet, 2> writes{};
-            uint32_t writeCount = 0;
-
-            if (geometry.vertex.buffer != VK_NULL_HANDLE && geometry.vertex.size > 0) {
-                geometry.descriptor.vertex.buffer = geometry.vertex.buffer;
-                geometry.descriptor.vertex.offset = geometry.vertex.offset;
-                geometry.descriptor.vertex.range = geometry.vertex.size;
-
-                VkWriteDescriptorSet &write = writes[writeCount++];
-                write.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-                write.dstSet = geometry.descriptor.set;
-                write.dstBinding = 0;
-                write.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-                write.descriptorCount = 1;
-                write.pBufferInfo = &geometry.descriptor.vertex;
-            } else {
-                geometry.descriptor.vertex = {};
-            }
-
-            if (geometry.index.buffer != VK_NULL_HANDLE && geometry.index.size > 0) {
-                geometry.descriptor.index.buffer = geometry.index.buffer;
-                geometry.descriptor.index.offset = geometry.index.offset;
-                geometry.descriptor.index.range = geometry.index.size;
-
-                VkWriteDescriptorSet &write = writes[writeCount++];
-                write.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-                write.dstSet = geometry.descriptor.set;
-                write.dstBinding = 1;
-                write.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-                write.descriptorCount = 1;
-                write.pBufferInfo = &geometry.descriptor.index;
-            } else {
-                geometry.descriptor.index = {};
-            }
-
-            if (writeCount > 0) {
-                vkUpdateDescriptorSets(device_, writeCount, writes.data(), 0, nullptr);
-            }
-        }
-    }
-
-    if (hasVertexData) {
-        geometry.vertexStaging.clear();
-    }
-    if (hasIndexData) {
-        geometry.indexStaging.clear();
-    }
-
-    geometry.uploaded = true;
-    return true;
-}
-
-void VulkanRenderer::destroyMeshGeometry(ModelRecord::MeshGeometry &geometry) {
-    if (geometry.descriptor.set != VK_NULL_HANDLE && descriptorPool_ != VK_NULL_HANDLE && device_ != VK_NULL_HANDLE) {
-        vkFreeDescriptorSets(device_, descriptorPool_, 1, &geometry.descriptor.set);
-    }
-    geometry.descriptor.set = VK_NULL_HANDLE;
-    geometry.descriptor.vertex = {};
-    geometry.descriptor.index = {};
-
-    destroyBuffer(geometry.vertex);
-    destroyBuffer(geometry.index);
-
-    geometry.vertexStaging.clear();
-    geometry.indexStaging.clear();
-    geometry.vertexCount = 0;
-    geometry.indexCount = 0;
-    geometry.indexType = VK_INDEX_TYPE_UINT16;
-    geometry.uploaded = false;
-}
-
-void VulkanRenderer::destroyModelRecord(ModelRecord &record) {
-    for (auto &geometry : record.meshGeometry) {
-        destroyMeshGeometry(geometry);
-    }
-    record.meshGeometry.clear();
-}
-
-void VulkanRenderer::destroyAllModelGeometry() {
-    for (auto &entry : models_) {
-        destroyModelRecord(entry.second);
-    }
-}
-
-VulkanRenderer::PipelineKind VulkanRenderer::selectPipelineForEntity(const entity_t &ent) const {
-    constexpr uint32_t kInlineMask = 1u << 31;
-    if ((ent.model & kInlineMask) != 0u) {
-        return PipelineKind::InlineBsp;
-    }
-    if (ent.flags & RF_WEAPONMODEL) {
-        return PipelineKind::Weapon;
-    }
-
-    const ModelRecord *record = findModelRecord(ent.model);
-    std::string_view extension = classifyModelName(record);
-    if (extension == ".sp2" || extension == ".spr" || extension == ".sprite") {
-        return PipelineKind::Sprite;
-    }
-
-    return PipelineKind::Alias;
-}
-
-void VulkanRenderer::classifyEntities(const refdef_t &fd) {
-    frameQueues_.clear();
-
-    static cvar_t *drawEntities = Cvar_FindVar("gl_drawentities");
-    if (drawEntities && drawEntities->integer == 0) {
-        return;
-    }
-
-    if (!fd.entities || fd.num_entities <= 0) {
-        return;
-    }
-
-    static cvar_t *drawOrder = Cvar_FindVar("gl_draworder");
-    float drawOrderThreshold = 0.5f;
-    if (drawOrder) {
-        drawOrderThreshold = drawOrder->value;
-    }
-
-    for (int i = 0; i < fd.num_entities; ++i) {
-        const entity_t *ent = &fd.entities[i];
-        if (ent->flags & RF_BEAM) {
-            if (ent->frame) {
-                frameQueues_.beams.push_back(ent);
-            }
-            continue;
-        }
-
-        if (ent->flags & RF_FLARE) {
-            frameQueues_.flares.push_back(ent);
-            continue;
-        }
-
-        constexpr uint32_t kInlineMask = 1u << 31;
-        if ((ent->model & kInlineMask) != 0u) {
-            frameQueues_.bmodels.push_back(ent);
-            continue;
-        }
-
-        if (!(ent->flags & RF_TRANSLUCENT)) {
-            frameQueues_.opaque.push_back(ent);
-            continue;
-        }
-
-        float alpha = ent->alpha;
-        if (!std::isfinite(alpha)) {
-            alpha = 1.0f;
-        }
-
-        if ((ent->flags & RF_WEAPONMODEL) || alpha <= drawOrderThreshold) {
-            frameQueues_.alphaFront.push_back(ent);
-            continue;
-        }
-
-        frameQueues_.alphaBack.push_back(ent);
-    }
-}
-
-void VulkanRenderer::sortTransparentQueues(const refdef_t &fd) {
-    if (frameQueues_.alphaBack.empty() && frameQueues_.alphaFront.empty()) {
-        return;
-    }
-
-    const vec3_t &viewOrg = fd.vieworg;
-    auto distanceSquared = [&](const entity_t *ent) {
-        if (!ent) {
-            return std::numeric_limits<float>::infinity();
-        }
-
-        float dx = ent->origin[0] - viewOrg[0];
-        float dy = ent->origin[1] - viewOrg[1];
-        float dz = ent->origin[2] - viewOrg[2];
-        float dist = dx * dx + dy * dy + dz * dz;
-        if (!std::isfinite(dist)) {
-            return std::numeric_limits<float>::infinity();
-        }
-        return dist;
-    };
-
-    auto sortByDepth = [&](std::vector<const entity_t *> &queue) {
-        if (queue.size() < 2) {
-            return;
-        }
-
-        std::stable_sort(queue.begin(), queue.end(), [&](const entity_t *lhs, const entity_t *rhs) {
-            float left = distanceSquared(lhs);
-            float right = distanceSquared(rhs);
-            if (left == right) {
-                return lhs < rhs;
-            }
-            return left < right;
-        });
-    };
-
-    sortByDepth(frameQueues_.alphaBack);
-    sortByDepth(frameQueues_.alphaFront);
-}
-
-void VulkanRenderer::buildEffectBuffers(const refdef_t &fd) {
-    framePrimitives_.clear();
-
-    for (const entity_t *beamEntity : frameQueues_.beams) {
-        BeamPrimitive primitive{};
-        primitive.start = toArray(beamEntity->origin);
-        primitive.end = toArray(beamEntity->oldorigin);
-        primitive.radius = static_cast<float>(beamEntity->frame);
-        primitive.color = beamEntity->rgba;
-        framePrimitives_.beams.push_back(primitive);
-    }
-    frameStats_.beams = framePrimitives_.beams.size();
-
-    if (fd.particles && fd.num_particles > 0) {
-        framePrimitives_.particles.reserve(static_cast<size_t>(fd.num_particles));
-        for (int i = 0; i < fd.num_particles; ++i) {
-            const particle_t &particle = fd.particles[i];
-            ParticleBillboard billboard{};
-            billboard.origin = toArray(particle.origin);
-            billboard.scale = particle.scale;
-            billboard.alpha = particle.alpha;
-            if (particle.color != -1) {
-                billboard.color.u32 = d_8to24table[particle.color & 0xff];
-            } else {
-                billboard.color = particle.rgba;
-            }
-            framePrimitives_.particles.push_back(billboard);
-        }
-    }
-    frameStats_.particles = framePrimitives_.particles.size();
-
-    for (const entity_t *flareEntity : frameQueues_.flares) {
-        FlarePrimitive flare{};
-        flare.origin = toArray(flareEntity->origin);
-        flare.scale = std::max(0.0f, flareEntity->alpha);
-        flare.color = flareEntity->rgba;
-        framePrimitives_.flares.push_back(flare);
-    }
-    frameStats_.flares = framePrimitives_.flares.size();
-
-    frameStats_.debugLines = framePrimitives_.debugLines.size();
-}
-
-VulkanRenderer::ViewParameters VulkanRenderer::computeViewParameters(const refdef_t &fd) const {
-    ViewParameters params{};
-    vec3_t axis[3]{};
-    AnglesToAxis(fd.viewangles, axis);
-    for (int i = 0; i < 3; ++i) {
-        for (int j = 0; j < 3; ++j) {
-            params.axis[i][j] = axis[i][j];
-        }
-    }
-    params.origin = toArray(fd.vieworg);
-    return params;
-}
-
-void VulkanRenderer::streamBeamPrimitives(const ViewParameters &view, bool cylindricalStyle) {
-    effectStreams_.beamVertices.clear();
-    effectStreams_.beamIndices.clear();
-
-    if (framePrimitives_.beams.empty()) {
-        return;
-    }
-
-    vec3_t viewAxis[3]{};
-    for (int i = 0; i < 3; ++i) {
-        arrayToVec3(view.axis[i], viewAxis[i]);
-    }
-    vec3_t viewOrigin{};
-    arrayToVec3(view.origin, viewOrigin);
-
-    for (const BeamPrimitive &beam : framePrimitives_.beams) {
-        vec3_t start{};
-        vec3_t end{};
-        arrayToVec3(beam.start, start);
-        arrayToVec3(beam.end, end);
-
-        float radius = std::abs(beam.radius);
-        if (!std::isfinite(radius) || radius <= 0.0f) {
-            continue;
-        }
-
-        float widthScale = cylindricalStyle ? 0.5f : 1.2f;
-        float width = radius * widthScale;
-        if (!std::isfinite(width) || width <= 0.0f) {
-            continue;
-        }
-
-        const std::array<float, 4> color = toColorArray(beam.color);
-
-        vec3_t direction{};
-        VectorSubtract(end, start, direction);
-        if (VectorNormalize(direction) <= 0.0f) {
-            continue;
-        }
-
-        if (cylindricalStyle) {
-            size_t requiredVertices = static_cast<size_t>(kBeamCylinderSides) * 2;
-            if (effectStreams_.beamVertices.size() + requiredVertices > std::numeric_limits<uint16_t>::max()) {
-                continue;
-            }
-
-            uint16_t baseIndex = static_cast<uint16_t>(effectStreams_.beamVertices.size());
-
-            vec3_t right{};
-            vec3_t up{};
-            MakeNormalVectors(direction, right, up);
-            VectorScale(right, width, right);
-
-            for (int i = 0; i < kBeamCylinderSides; ++i) {
-                float angle = (360.0f / static_cast<float>(kBeamCylinderSides)) * static_cast<float>(i);
-                vec3_t offset{};
-                RotatePointAroundVector(offset, direction, right, angle);
-
-                vec3_t startVertex{};
-                VectorAdd(start, offset, startVertex);
-                EffectVertexStreams::BeamVertex startBeam{};
-                startBeam.position = toArray(startVertex);
-                startBeam.uv = { 0.0f, 0.0f };
-                startBeam.color = color;
-                effectStreams_.beamVertices.emplace_back(startBeam);
-
-                vec3_t endVertex{};
-                VectorAdd(end, offset, endVertex);
-                EffectVertexStreams::BeamVertex endBeam{};
-                endBeam.position = toArray(endVertex);
-                endBeam.uv = { 0.0f, 1.0f };
-                endBeam.color = color;
-                effectStreams_.beamVertices.emplace_back(endBeam);
-            }
-
-            for (int i = 0; i < kBeamCylinderSides; ++i) {
-                uint16_t current = baseIndex + static_cast<uint16_t>(i * 2);
-                uint16_t next = baseIndex + static_cast<uint16_t>(((i + 1) % kBeamCylinderSides) * 2);
-                uint16_t currentTop = current + 1;
-                uint16_t nextTop = next + 1;
-
-                effectStreams_.beamIndices.push_back(current);
-                effectStreams_.beamIndices.push_back(currentTop);
-                effectStreams_.beamIndices.push_back(nextTop);
-
-                effectStreams_.beamIndices.push_back(current);
-                effectStreams_.beamIndices.push_back(nextTop);
-                effectStreams_.beamIndices.push_back(next);
-            }
-        } else {
-            if (effectStreams_.beamVertices.size() + 4 > std::numeric_limits<uint16_t>::max()) {
-                continue;
-            }
-
-            vec3_t viewerToStart{};
-            VectorSubtract(viewOrigin, start, viewerToStart);
-
-            vec3_t right{};
-            CrossProduct(direction, viewerToStart, right);
-            if (VectorNormalize(right) <= 0.0f) {
-                vec3_t tmp{};
-                MakeNormalVectors(direction, right, tmp);
-            }
-            VectorScale(right, width, right);
-
-            vec3_t v0{};
-            vec3_t v1{};
-            vec3_t v2{};
-            vec3_t v3{};
-            VectorAdd(start, right, v0);
-            VectorSubtract(start, right, v1);
-            VectorSubtract(end, right, v2);
-            VectorAdd(end, right, v3);
-
-            uint16_t baseIndex = static_cast<uint16_t>(effectStreams_.beamVertices.size());
-
-            EffectVertexStreams::BeamVertex vertices[4]{};
-            vertices[0].position = toArray(v0);
-            vertices[0].uv = { 0.0f, 0.0f };
-            vertices[0].color = color;
-
-            vertices[1].position = toArray(v1);
-            vertices[1].uv = { 1.0f, 0.0f };
-            vertices[1].color = color;
-
-            vertices[2].position = toArray(v2);
-            vertices[2].uv = { 1.0f, 1.0f };
-            vertices[2].color = color;
-
-            vertices[3].position = toArray(v3);
-            vertices[3].uv = { 0.0f, 1.0f };
-            vertices[3].color = color;
-
-            effectStreams_.beamVertices.insert(effectStreams_.beamVertices.end(), std::begin(vertices), std::end(vertices));
-
-            effectStreams_.beamIndices.push_back(baseIndex + 0);
-            effectStreams_.beamIndices.push_back(baseIndex + 2);
-            effectStreams_.beamIndices.push_back(baseIndex + 3);
-            effectStreams_.beamIndices.push_back(baseIndex + 0);
-            effectStreams_.beamIndices.push_back(baseIndex + 1);
-            effectStreams_.beamIndices.push_back(baseIndex + 2);
-        }
-    }
-}
-
-void VulkanRenderer::streamParticlePrimitives(const ViewParameters &view, bool additiveBlend) {
-    (void)additiveBlend;
-
-    effectStreams_.particleVertices.clear();
-
-    if (framePrimitives_.particles.empty()) {
-        return;
-    }
-
-    vec3_t viewAxis[3]{};
-    for (int i = 0; i < 3; ++i) {
-        arrayToVec3(view.axis[i], viewAxis[i]);
-    }
-    vec3_t viewOrigin{};
-    arrayToVec3(view.origin, viewOrigin);
-
-    float partScale = (gl_partscale) ? gl_partscale->value : 1.0f;
-
-    for (const ParticleBillboard &billboard : framePrimitives_.particles) {
-        vec3_t origin{};
-        arrayToVec3(billboard.origin, origin);
-
-        vec3_t toParticle{};
-        VectorSubtract(origin, viewOrigin, toParticle);
-        float dist = DotProduct(toParticle, viewAxis[0]);
-
-        float scale = 1.0f;
-        if (dist > kParticleDistanceBias) {
-            scale += dist * kParticleDistanceScale;
-        }
-        scale *= partScale * billboard.scale;
-        if (!std::isfinite(scale) || scale <= 0.0f) {
-            continue;
-        }
-
-        float scale2 = scale * kParticleScale;
-
-        vec3_t vertex0{};
-        VectorMA(origin, scale2, viewAxis[1], vertex0);
-        VectorMA(vertex0, -scale2, viewAxis[2], vertex0);
-
-        vec3_t vertex1{};
-        VectorMA(vertex0, scale, viewAxis[2], vertex1);
-
-        vec3_t vertex2{};
-        VectorMA(vertex0, -scale, viewAxis[1], vertex2);
-
-        const std::array<float, 4> color = toColorArray(billboard.color, billboard.alpha);
-
-        EffectVertexStreams::BillboardVertex v0{};
-        v0.position = toArray(vertex0);
-        v0.uv = { 0.0f, 0.0f };
-        v0.color = color;
-
-        EffectVertexStreams::BillboardVertex v1{};
-        v1.position = toArray(vertex1);
-        v1.uv = { 0.0f, kParticleSize };
-        v1.color = color;
-
-        EffectVertexStreams::BillboardVertex v2{};
-        v2.position = toArray(vertex2);
-        v2.uv = { kParticleSize, 0.0f };
-        v2.color = color;
-
-        effectStreams_.particleVertices.push_back(v0);
-        effectStreams_.particleVertices.push_back(v1);
-        effectStreams_.particleVertices.push_back(v2);
-    }
-}
-
-void VulkanRenderer::streamFlarePrimitives(const ViewParameters &view) {
-    effectStreams_.flareVertices.clear();
-    effectStreams_.flareIndices.clear();
-
-    if (framePrimitives_.flares.empty()) {
-        return;
-    }
-
-    vec3_t viewAxis[3]{};
-    for (int i = 0; i < 3; ++i) {
-        arrayToVec3(view.axis[i], viewAxis[i]);
-    }
-    vec3_t viewOrigin{};
-    arrayToVec3(view.origin, viewOrigin);
-
-    for (const FlarePrimitive &flare : framePrimitives_.flares) {
-        if (flare.scale <= 0.0f) {
-            continue;
-        }
-
-        vec3_t origin{};
-        arrayToVec3(flare.origin, origin);
-
-        vec3_t dir{};
-        VectorSubtract(origin, viewOrigin, dir);
-        float dist = DotProduct(dir, viewAxis[0]);
-
-        float scale = 2.5f;
-        if (dist > kParticleDistanceBias) {
-            scale += dist * kParticleDistanceScale;
-        }
-        scale *= flare.scale;
-        if (!std::isfinite(scale) || scale <= 0.0f) {
-            continue;
-        }
-
-        if (effectStreams_.flareVertices.size() + 4 > std::numeric_limits<uint16_t>::max()) {
-            continue;
-        }
-
-        vec3_t left{};
-        vec3_t right{};
-        vec3_t up{};
-        vec3_t down{};
-        VectorScale(viewAxis[1], scale, left);
-        VectorScale(viewAxis[1], -scale, right);
-        VectorScale(viewAxis[2], scale, up);
-        VectorScale(viewAxis[2], -scale, down);
-
-        vec3_t v0{};
-        vec3_t v1{};
-        vec3_t v2{};
-        vec3_t v3{};
-
-        vec3_t temp{};
-        VectorAdd(origin, down, temp);
-        VectorAdd(temp, left, v0);
-
-        VectorAdd(origin, up, temp);
-        VectorAdd(temp, left, v1);
-
-        VectorAdd(origin, down, temp);
-        VectorAdd(temp, right, v2);
-
-        VectorAdd(origin, up, temp);
-        VectorAdd(temp, right, v3);
-
-        uint16_t baseIndex = static_cast<uint16_t>(effectStreams_.flareVertices.size());
-
-        const std::array<float, 4> color = toColorArray(flare.color);
-
-        EffectVertexStreams::BillboardVertex vertices[4]{};
-        vertices[0].position = toArray(v0);
-        vertices[0].uv = { 0.0f, 1.0f };
-        vertices[0].color = color;
-
-        vertices[1].position = toArray(v1);
-        vertices[1].uv = { 0.0f, 0.0f };
-        vertices[1].color = color;
-
-        vertices[2].position = toArray(v2);
-        vertices[2].uv = { 1.0f, 1.0f };
-        vertices[2].color = color;
-
-        vertices[3].position = toArray(v3);
-        vertices[3].uv = { 1.0f, 0.0f };
-        vertices[3].color = color;
-
-        effectStreams_.flareVertices.insert(effectStreams_.flareVertices.end(), std::begin(vertices), std::end(vertices));
-
-        effectStreams_.flareIndices.push_back(baseIndex + 0);
-        effectStreams_.flareIndices.push_back(baseIndex + 1);
-        effectStreams_.flareIndices.push_back(baseIndex + 2);
-        effectStreams_.flareIndices.push_back(baseIndex + 2);
-        effectStreams_.flareIndices.push_back(baseIndex + 1);
-        effectStreams_.flareIndices.push_back(baseIndex + 3);
-    }
-}
-
-void VulkanRenderer::streamDebugLinePrimitives() {
-    effectStreams_.debugLinesDepth.clear();
-    effectStreams_.debugLinesNoDepth.clear();
-
-    if (framePrimitives_.debugLines.empty()) {
-        return;
-    }
-
-    for (const DebugLinePrimitive &line : framePrimitives_.debugLines) {
-        const std::array<float, 4> color = toColorArray(line.color);
-
-        EffectVertexStreams::DebugLineVertex start{};
-        start.position = line.start;
-        start.color = color;
-
-        EffectVertexStreams::DebugLineVertex end{};
-        end.position = line.end;
-        end.color = color;
-
-        if (line.depthTest) {
-            effectStreams_.debugLinesDepth.push_back(start);
-            effectStreams_.debugLinesDepth.push_back(end);
-        } else {
-            effectStreams_.debugLinesNoDepth.push_back(start);
-            effectStreams_.debugLinesNoDepth.push_back(end);
-        }
-    }
-}
-
-void VulkanRenderer::submit2DDraw(const draw2d::Submission &submission) {
-    if (!submission.vertices || !submission.indices) {
-        return;
-    }
-
-    if (submission.vertexCount == 0 || submission.indexCount == 0) {
-        return;
-    }
-
-    if (!initialized_ || device_ == VK_NULL_HANDLE || inFlightFrames_.empty()) {
-        return;
-    }
-
-    if (currentFrameIndex_ >= inFlightFrames_.size()) {
-        return;
-    }
-
-    InFlightFrame &frame = inFlightFrames_[currentFrameIndex_];
-    if (frame.commandBuffer == VK_NULL_HANDLE || !frame.hasImage) {
-        return;
-    }
-
-    VkCommandBuffer commandBuffer = frame.commandBuffer;
-
-    Draw2DBatch batch{};
-    batch.vertexCount = submission.vertexCount;
-    batch.indexCount = submission.indexCount;
-    batch.texture = submission.texture;
-
-    const VkDeviceSize vertexSize = static_cast<VkDeviceSize>(submission.vertexCount * sizeof(draw2d::Vertex));
-    const VkDeviceSize indexSize = static_cast<VkDeviceSize>(submission.indexCount * sizeof(uint16_t));
-
-    if (vertexSize == 0 || indexSize == 0) {
-        return;
-    }
-
-    if (!createBuffer(batch.vertex,
-                      vertexSize,
-                      VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-                      VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT)) {
-        return;
-    }
-
-    void *mappedVertices = nullptr;
-    if (vkMapMemory(device_, batch.vertex.memory, batch.vertex.offset, batch.vertex.size, 0, &mappedVertices) != VK_SUCCESS ||
-        mappedVertices == nullptr) {
-        destroyBuffer(batch.vertex);
-        return;
-    }
-    std::memcpy(mappedVertices, submission.vertices, static_cast<size_t>(vertexSize));
-    vkUnmapMemory(device_, batch.vertex.memory);
-
-    if (!createBuffer(batch.index,
-                      indexSize,
-                      VK_BUFFER_USAGE_INDEX_BUFFER_BIT,
-                      VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT)) {
-        destroyBuffer(batch.vertex);
-        return;
-    }
-
-    void *mappedIndices = nullptr;
-    if (vkMapMemory(device_, batch.index.memory, batch.index.offset, batch.index.size, 0, &mappedIndices) != VK_SUCCESS ||
-        mappedIndices == nullptr) {
-        destroyBuffer(batch.vertex);
-        destroyBuffer(batch.index);
-        return;
-    }
-    std::memcpy(mappedIndices, submission.indices, static_cast<size_t>(indexSize));
-    vkUnmapMemory(device_, batch.index.memory);
-
-    batch.descriptor.vertex.buffer = batch.vertex.buffer;
-    batch.descriptor.vertex.offset = batch.vertex.offset;
-    batch.descriptor.vertex.range = batch.vertex.size;
-    batch.descriptor.index.buffer = batch.index.buffer;
-    batch.descriptor.index.offset = batch.index.offset;
-    batch.descriptor.index.range = batch.index.size;
-
-    if (descriptorPool_ != VK_NULL_HANDLE && createModelDescriptorResources()) {
-        if (batch.descriptor.set == VK_NULL_HANDLE) {
-            VkDescriptorSetAllocateInfo allocInfo{};
-            allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-            allocInfo.descriptorPool = descriptorPool_;
-            allocInfo.descriptorSetCount = 1;
-            allocInfo.pSetLayouts = &modelDescriptorSetLayout_;
-            VkResult allocResult = vkAllocateDescriptorSets(device_, &allocInfo, &batch.descriptor.set);
-            if (allocResult != VK_SUCCESS) {
-                batch.descriptor.set = VK_NULL_HANDLE;
-            }
-        }
-
-        if (batch.descriptor.set != VK_NULL_HANDLE) {
-            std::array<VkWriteDescriptorSet, 2> writes{};
-
-            VkWriteDescriptorSet &vertexWrite = writes[0];
-            vertexWrite.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            vertexWrite.dstSet = batch.descriptor.set;
-            vertexWrite.dstBinding = 0;
-            vertexWrite.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-            vertexWrite.descriptorCount = 1;
-            vertexWrite.pBufferInfo = &batch.descriptor.vertex;
-
-            VkWriteDescriptorSet &indexWrite = writes[1];
-            indexWrite.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            indexWrite.dstSet = batch.descriptor.set;
-            indexWrite.dstBinding = 1;
-            indexWrite.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-            indexWrite.descriptorCount = 1;
-            indexWrite.pBufferInfo = &batch.descriptor.index;
-
-            vkUpdateDescriptorSets(device_, static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
-        }
-    }
-
-    if (batch.vertex.buffer != VK_NULL_HANDLE) {
-        VkBuffer buffers[] = { batch.vertex.buffer };
-        VkDeviceSize offsets[] = { batch.vertex.offset };
-        vkCmdBindVertexBuffers(commandBuffer, 0, 1, buffers, offsets);
-    }
-
-    if (batch.index.buffer != VK_NULL_HANDLE) {
-        vkCmdBindIndexBuffer(commandBuffer, batch.index.buffer, batch.index.offset, batch.indexType);
-    }
-
-    if (batch.descriptor.set != VK_NULL_HANDLE && modelPipelineLayout_ != VK_NULL_HANDLE) {
-        vkCmdBindDescriptorSets(commandBuffer,
-                                VK_PIPELINE_BIND_POINT_GRAPHICS,
-                                modelPipelineLayout_,
-                                0,
-                                1,
-                                &batch.descriptor.set,
-                                0,
-                                nullptr);
-    }
-
-    vkCmdDrawIndexed(commandBuffer, static_cast<uint32_t>(submission.indexCount), 1, 0, 0, 0);
-
-    frame2DBatches_.push_back(batch);
-
-    const PipelineDesc &pipeline = ensurePipeline(buildPipelineKey(PipelineKind::Draw2D));
-    recordDrawCall(pipeline, "draw2d", submission.indexCount / 3);
-}
-
-bool VulkanRenderer::canSubmit2D() const {
-    return frameActive_ && draw2d::isActive();
-}
-
-qhandle_t VulkanRenderer::ensureWhiteTexture() {
-    if (whiteTextureHandle_ != 0) {
-        if (auto it = images_.find(whiteTextureHandle_); it != images_.end() && it->second.image != VK_NULL_HANDLE) {
-            return whiteTextureHandle_;
-        }
-        whiteTextureHandle_ = 0;
-    }
-
-    qhandle_t handle = registerResource(imageLookup_, "__vk_white");
-    whiteTextureHandle_ = handle;
-
-    ImageRecord record{};
-    record.handle = handle;
-    record.name = "__vk_white";
-    record.type = IT_PIC;
-    record.flags = static_cast<imageflags_t>(IF_SPECIAL | IF_PERMANENT);
-    record.width = 1;
-    record.height = 1;
-    record.uploadWidth = 1;
-    record.uploadHeight = 1;
-    record.transparent = false;
-    record.registrationSequence = r_registration_sequence;
-
-    const uint8_t whitePixel[4] = { 0xFF, 0xFF, 0xFF, 0xFF };
-    if (!ensureTextureResources(record, whitePixel, sizeof(whitePixel), 1, 1, VK_FORMAT_R8G8B8A8_UNORM)) {
-        imageLookup_.erase(record.name);
-        whiteTextureHandle_ = 0;
-        return 0;
-    }
-
-    images_[handle] = record;
-    imageLookup_[record.name] = handle;
-    return handle;
-}
-
-qhandle_t VulkanRenderer::ensureRawTexture() {
-    if (rawTextureHandle_ != 0) {
-        if (auto it = images_.find(rawTextureHandle_); it != images_.end()) {
-            ImageRecord &record = it->second;
-            record.width = rawPic_.width;
-            record.height = rawPic_.height;
-            record.uploadWidth = rawPic_.width;
-            record.uploadHeight = rawPic_.height;
-            record.registrationSequence = r_registration_sequence;
-
-            if (!rawPic_.pixels.empty()) {
-                const uint8_t *pixels = reinterpret_cast<const uint8_t *>(rawPic_.pixels.data());
-                size_t size = rawPic_.pixels.size() * sizeof(uint32_t);
-                uint32_t width = static_cast<uint32_t>(std::max(1, rawPic_.width));
-                uint32_t height = static_cast<uint32_t>(std::max(1, rawPic_.height));
-                ensureTextureResources(record, pixels, size, width, height, VK_FORMAT_R8G8B8A8_UNORM);
-            }
-            return rawTextureHandle_;
-        }
-        rawTextureHandle_ = 0;
-    }
-
-    if (rawPic_.pixels.empty() || rawPic_.width <= 0 || rawPic_.height <= 0) {
-        return 0;
-    }
-
-    qhandle_t handle = registerResource(imageLookup_, "__vk_raw");
-    rawTextureHandle_ = handle;
-
-    ImageRecord record{};
-    record.handle = handle;
-    record.name = "__vk_raw";
-    record.type = IT_PIC;
-    record.flags = IF_SPECIAL;
-    record.width = rawPic_.width;
-    record.height = rawPic_.height;
-    record.uploadWidth = rawPic_.width;
-    record.uploadHeight = rawPic_.height;
-    record.transparent = true;
-    record.registrationSequence = r_registration_sequence;
-
-    const uint8_t *pixels = reinterpret_cast<const uint8_t *>(rawPic_.pixels.data());
-    size_t size = rawPic_.pixels.size() * sizeof(uint32_t);
-    uint32_t width = static_cast<uint32_t>(std::max(1, rawPic_.width));
-    uint32_t height = static_cast<uint32_t>(std::max(1, rawPic_.height));
-
-    if (!ensureTextureResources(record, pixels, size, width, height, VK_FORMAT_R8G8B8A8_UNORM)) {
-        imageLookup_.erase(record.name);
-        rawTextureHandle_ = 0;
-        return 0;
-    }
-
-    images_[handle] = record;
-    imageLookup_[record.name] = handle;
-    return handle;
-}
-
-const std::vector<std::string> &VulkanRenderer::platformInstanceExtensions() const {
-    return platformInstanceExtensions_;
-}
-
-VkSurfaceKHR VulkanRenderer::platformSurface() const {
-    return platformSurface_;
-}
-
-bool VulkanRenderer::createPlatformSurface(VkInstance instance, const VkAllocationCallbacks *allocator) {
-    if (!platformHooks_.createSurface) {
-        Com_Printf("refresh-vk: platform does not support Vulkan surface creation.\n");
-        return false;
-    }
-
-    destroyPlatformSurface(VK_NULL_HANDLE, allocator);
-
-    VkSurfaceKHR surface = VK_NULL_HANDLE;
-    VkResult result = platformHooks_.createSurface(instance, allocator, &surface);
-    if (result != VK_SUCCESS || surface == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create Vulkan surface (VkResult %d).\n", static_cast<int>(result));
-        return false;
-    }
-
-    platformSurface_ = surface;
-    platformInstance_ = instance;
-    return true;
-}
-
-void VulkanRenderer::destroyPlatformSurface(VkInstance instance, const VkAllocationCallbacks *allocator) {
-    if (platformSurface_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    VkInstance destroyInstance = instance;
-    if (destroyInstance == VK_NULL_HANDLE) {
-        destroyInstance = platformInstance_;
-    }
-
-    if (platformHooks_.destroySurface && destroyInstance != VK_NULL_HANDLE) {
-        platformHooks_.destroySurface(destroyInstance, platformSurface_, allocator);
-    }
-
-    platformSurface_ = VK_NULL_HANDLE;
-    platformInstance_ = VK_NULL_HANDLE;
-}
-
-void VulkanRenderer::initializePlatformHooks() {
-    platformHooks_ = {};
-    platformInstance_ = VK_NULL_HANDLE;
-    platformSurface_ = VK_NULL_HANDLE;
-    supportsDebugUtils_ = false;
-
-    if (vid) {
-        platformHooks_.getInstanceExtensions = vid->vk.get_instance_extensions;
-        platformHooks_.createSurface = vid->vk.create_surface;
-        platformHooks_.destroySurface = vid->vk.destroy_surface;
-    }
-
-    collectPlatformInstanceExtensions();
-}
-
-void VulkanRenderer::collectPlatformInstanceExtensions() {
-    platformInstanceExtensions_.clear();
-
-    std::unordered_set<std::string> seen;
-    auto addExtension = [&](const char *name) {
-        if (!name || !*name) {
-            return;
-        }
-        if (std::strcmp(name, VK_EXT_DEBUG_UTILS_EXTENSION_NAME) == 0) {
-            supportsDebugUtils_ = true;
-        }
-        if (seen.insert(name).second) {
-            platformInstanceExtensions_.emplace_back(name);
-        }
-    };
-
-    addExtension(VK_KHR_SURFACE_EXTENSION_NAME);
-
-    if (!platformHooks_.getInstanceExtensions) {
-        return;
-    }
-
-    uint32_t count = 0;
-    const char *const *extensions = platformHooks_.getInstanceExtensions(&count);
-    if (!extensions || count == 0) {
-        return;
-    }
-
-    platformInstanceExtensions_.reserve(platformInstanceExtensions_.size() + count);
-    for (uint32_t i = 0; i < count; ++i) {
-        addExtension(extensions[i]);
-    }
-}
-
-bool VulkanRenderer::createInstance() {
-    if (instance_ != VK_NULL_HANDLE) {
-        return true;
-    }
-
-    std::vector<const char *> extensions;
-    extensions.reserve(platformInstanceExtensions_.size());
-    for (const std::string &extension : platformInstanceExtensions_) {
-        extensions.push_back(extension.c_str());
-    }
-
-    VkApplicationInfo appInfo{};
-    appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-    appInfo.pApplicationName = "WORR";
-    appInfo.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
-    appInfo.pEngineName = "refresh-vk";
-    appInfo.engineVersion = VK_MAKE_VERSION(1, 0, 0);
-    appInfo.apiVersion = VK_API_VERSION_1_1;
-
-    VkInstanceCreateInfo instanceInfo{};
-    instanceInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-    instanceInfo.pApplicationInfo = &appInfo;
-    instanceInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
-    instanceInfo.ppEnabledExtensionNames = extensions.empty() ? nullptr : extensions.data();
-
-    VkResult result = vkCreateInstance(&instanceInfo, nullptr, &instance_);
-    if (result != VK_SUCCESS || instance_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create Vulkan instance (VkResult %d).\n", static_cast<int>(result));
-        instance_ = VK_NULL_HANDLE;
-        return false;
-    }
-
-    return true;
-}
-
-void VulkanRenderer::destroyInstance() {
-    if (instance_ != VK_NULL_HANDLE) {
-        vkDestroyInstance(instance_, nullptr);
-        instance_ = VK_NULL_HANDLE;
-    }
-}
-
-VulkanRenderer::QueueFamilyIndices VulkanRenderer::queryQueueFamilies(VkPhysicalDevice device) const {
-    QueueFamilyIndices indices{};
-
-    uint32_t count = 0;
-    vkGetPhysicalDeviceQueueFamilyProperties(device, &count, nullptr);
-    std::vector<VkQueueFamilyProperties> properties(count);
-    if (count > 0) {
-        vkGetPhysicalDeviceQueueFamilyProperties(device, &count, properties.data());
-    }
-
-    for (uint32_t i = 0; i < count; ++i) {
-        if (!indices.graphics && (properties[i].queueFlags & VK_QUEUE_GRAPHICS_BIT)) {
-            indices.graphics = i;
-        }
-
-        if (!indices.present && platformSurface_ != VK_NULL_HANDLE) {
-            VkBool32 supported = VK_FALSE;
-            vkGetPhysicalDeviceSurfaceSupportKHR(device, i, platformSurface_, &supported);
-            if (supported) {
-                indices.present = i;
-            }
-        }
-
-        if (indices.complete()) {
-            break;
-        }
-    }
-
-    return indices;
-}
-
-VulkanRenderer::SwapchainSupportDetails VulkanRenderer::querySwapchainSupport(VkPhysicalDevice device) const {
-    SwapchainSupportDetails details{};
-    if (platformSurface_ == VK_NULL_HANDLE) {
-        return details;
-    }
-
-    vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, platformSurface_, &details.capabilities);
-
-    uint32_t formatCount = 0;
-    vkGetPhysicalDeviceSurfaceFormatsKHR(device, platformSurface_, &formatCount, nullptr);
-    if (formatCount > 0) {
-        details.formats.resize(formatCount);
-        vkGetPhysicalDeviceSurfaceFormatsKHR(device, platformSurface_, &formatCount, details.formats.data());
-    }
-
-    uint32_t presentCount = 0;
-    vkGetPhysicalDeviceSurfacePresentModesKHR(device, platformSurface_, &presentCount, nullptr);
-    if (presentCount > 0) {
-        details.presentModes.resize(presentCount);
-        vkGetPhysicalDeviceSurfacePresentModesKHR(device, platformSurface_, &presentCount, details.presentModes.data());
-    }
-
-    return details;
-}
-
-VkSurfaceFormatKHR VulkanRenderer::chooseSwapchainFormat(const std::vector<VkSurfaceFormatKHR> &formats) const {
-    for (const VkSurfaceFormatKHR &format : formats) {
-        if (format.format == VK_FORMAT_B8G8R8A8_UNORM && format.colorSpace == VK_COLOR_SPACE_SRGB_NONLINEAR_KHR) {
-            return format;
-        }
-    }
-
-    if (!formats.empty()) {
-        return formats.front();
-    }
-
-    VkSurfaceFormatKHR fallback{};
-    fallback.format = VK_FORMAT_B8G8R8A8_UNORM;
-    fallback.colorSpace = VK_COLOR_SPACE_SRGB_NONLINEAR_KHR;
-    return fallback;
-}
-
-VkPresentModeKHR VulkanRenderer::choosePresentMode(const std::vector<VkPresentModeKHR> &presentModes) const {
-    auto hasMode = [&](VkPresentModeKHR desired) {
-        return std::find(presentModes.begin(), presentModes.end(), desired) != presentModes.end();
-    };
-
-    bool wantVsync = swapInterval_ > 0;
-
-    if (!wantVsync) {
-        if (hasMode(VK_PRESENT_MODE_MAILBOX_KHR)) {
-            return VK_PRESENT_MODE_MAILBOX_KHR;
-        }
-        if (hasMode(VK_PRESENT_MODE_IMMEDIATE_KHR)) {
-            return VK_PRESENT_MODE_IMMEDIATE_KHR;
-        }
-        if (hasMode(VK_PRESENT_MODE_FIFO_RELAXED_KHR)) {
-            return VK_PRESENT_MODE_FIFO_RELAXED_KHR;
-        }
-    } else {
-        if (swapInterval_ > 1 && hasMode(VK_PRESENT_MODE_FIFO_RELAXED_KHR)) {
-            return VK_PRESENT_MODE_FIFO_RELAXED_KHR;
-        }
-        if (hasMode(VK_PRESENT_MODE_FIFO_KHR)) {
-            return VK_PRESENT_MODE_FIFO_KHR;
-        }
-        if (hasMode(VK_PRESENT_MODE_FIFO_RELAXED_KHR)) {
-            return VK_PRESENT_MODE_FIFO_RELAXED_KHR;
-        }
-    }
-
-    if (!presentModes.empty()) {
-        return presentModes.front();
-    }
-
-    return VK_PRESENT_MODE_FIFO_KHR;
-}
-
-VkExtent2D VulkanRenderer::chooseSwapchainExtent(const VkSurfaceCapabilitiesKHR &capabilities) const {
-    if (capabilities.currentExtent.width != std::numeric_limits<uint32_t>::max()) {
-        return capabilities.currentExtent;
-    }
-
-    VkExtent2D extent{};
-    extent.width = static_cast<uint32_t>(std::max(1, r_config.width));
-    extent.height = static_cast<uint32_t>(std::max(1, r_config.height));
-
-    extent.width = std::clamp(extent.width, capabilities.minImageExtent.width, capabilities.maxImageExtent.width);
-    extent.height = std::clamp(extent.height, capabilities.minImageExtent.height, capabilities.maxImageExtent.height);
-    return extent;
-}
-
-bool VulkanRenderer::pickPhysicalDevice() {
-    if (instance_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    memoryProperties_ = {};
-    uint32_t deviceCount = 0;
-    vkEnumeratePhysicalDevices(instance_, &deviceCount, nullptr);
-    if (deviceCount == 0) {
-        Com_Printf("refresh-vk: no Vulkan physical devices available.\n");
-        return false;
-    }
-
-    std::vector<VkPhysicalDevice> devices(deviceCount);
-    vkEnumeratePhysicalDevices(instance_, &deviceCount, devices.data());
-
-    for (VkPhysicalDevice candidate : devices) {
-        QueueFamilyIndices indices = queryQueueFamilies(candidate);
-        if (!indices.complete()) {
-            continue;
-        }
-
-        uint32_t extensionCount = 0;
-        vkEnumerateDeviceExtensionProperties(candidate, nullptr, &extensionCount, nullptr);
-        std::vector<VkExtensionProperties> extensions(extensionCount);
-        if (extensionCount > 0) {
-            vkEnumerateDeviceExtensionProperties(candidate, nullptr, &extensionCount, extensions.data());
-        }
-
-        bool hasSwapchain = false;
-        for (const VkExtensionProperties &prop : extensions) {
-            if (std::strcmp(prop.extensionName, VK_KHR_SWAPCHAIN_EXTENSION_NAME) == 0) {
-                hasSwapchain = true;
-                break;
-            }
-        }
-
-        if (!hasSwapchain) {
-            continue;
-        }
-
-        SwapchainSupportDetails support = querySwapchainSupport(candidate);
-        if (support.formats.empty() || support.presentModes.empty()) {
-            continue;
-        }
-
-        deviceExtensions_.clear();
-        deviceExtensions_.reserve(extensions.size());
-        for (const VkExtensionProperties &prop : extensions) {
-            deviceExtensions_.emplace_back(prop.extensionName);
-            if (std::strcmp(prop.extensionName, VK_EXT_DEBUG_MARKER_EXTENSION_NAME) == 0) {
-                supportsDebugUtils_ = true;
-            }
-        }
-
-        vkGetPhysicalDeviceProperties(candidate, &physicalDeviceProperties_);
-        vkGetPhysicalDeviceFeatures(candidate, &supportedFeatures_);
-
-        physicalDevice_ = candidate;
-        graphicsQueueFamily_ = indices.graphics.value();
-        presentQueueFamily_ = indices.present.value();
-        vkGetPhysicalDeviceMemoryProperties(physicalDevice_, &memoryProperties_);
-        return true;
-    }
-
-    Com_Printf("refresh-vk: failed to find suitable Vulkan physical device.\n");
-    return false;
-}
-
-bool VulkanRenderer::createLogicalDevice() {
-    if (physicalDevice_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    std::vector<VkDeviceQueueCreateInfo> queueInfos;
-    std::vector<uint32_t> uniqueFamilies;
-    uniqueFamilies.push_back(graphicsQueueFamily_);
-    if (presentQueueFamily_ != graphicsQueueFamily_) {
-        uniqueFamilies.push_back(presentQueueFamily_);
-    }
-
-    float priority = 1.0f;
-    for (uint32_t family : uniqueFamilies) {
-        VkDeviceQueueCreateInfo info{};
-        info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-        info.queueFamilyIndex = family;
-        info.queueCount = 1;
-        info.pQueuePriorities = &priority;
-        queueInfos.push_back(info);
-    }
-
-    VkPhysicalDeviceFeatures features{};
-
-    const char *extensions[] = { VK_KHR_SWAPCHAIN_EXTENSION_NAME };
-
-    VkDeviceCreateInfo deviceInfo{};
-    deviceInfo.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-    deviceInfo.queueCreateInfoCount = static_cast<uint32_t>(queueInfos.size());
-    deviceInfo.pQueueCreateInfos = queueInfos.data();
-    deviceInfo.pEnabledFeatures = &features;
-    deviceInfo.enabledExtensionCount = static_cast<uint32_t>(sizeof(extensions) / sizeof(extensions[0]));
-    deviceInfo.ppEnabledExtensionNames = extensions;
-
-    VkResult result = vkCreateDevice(physicalDevice_, &deviceInfo, nullptr, &device_);
-    if (result != VK_SUCCESS || device_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create logical device (VkResult %d).\n", static_cast<int>(result));
-        device_ = VK_NULL_HANDLE;
-        return false;
-    }
-
-    enabledFeatures_ = features;
-
-    vkGetDeviceQueue(device_, graphicsQueueFamily_, 0, &graphicsQueue_);
-    vkGetDeviceQueue(device_, presentQueueFamily_, 0, &presentQueue_);
-
-    return graphicsQueue_ != VK_NULL_HANDLE && presentQueue_ != VK_NULL_HANDLE;
-}
-
-bool VulkanRenderer::createCommandPool() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    if (commandPool_ != VK_NULL_HANDLE) {
-        return true;
-    }
-
-    VkCommandPoolCreateInfo info{};
-    info.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-    info.queueFamilyIndex = graphicsQueueFamily_;
-    info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-
-    VkResult result = vkCreateCommandPool(device_, &info, nullptr, &commandPool_);
-    if (result != VK_SUCCESS || commandPool_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create command pool (VkResult %d).\n", static_cast<int>(result));
-        commandPool_ = VK_NULL_HANDLE;
-        return false;
-    }
-
-    return true;
-}
-
-bool VulkanRenderer::createSwapchainResources(VkSwapchainKHR oldSwapchain) {
-    if (device_ == VK_NULL_HANDLE || platformSurface_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    SwapchainSupportDetails support = querySwapchainSupport(physicalDevice_);
-    if (support.formats.empty() || support.presentModes.empty()) {
-        Com_Printf("refresh-vk: swapchain support incomplete.\n");
-        return false;
-    }
-
-    VkSurfaceFormatKHR surfaceFormat = chooseSwapchainFormat(support.formats);
-    VkPresentModeKHR presentMode = choosePresentMode(support.presentModes);
-    VkExtent2D extent = chooseSwapchainExtent(support.capabilities);
-
-    uint32_t imageCount = support.capabilities.minImageCount + 1;
-    if (support.capabilities.maxImageCount > 0 && imageCount > support.capabilities.maxImageCount) {
-        imageCount = support.capabilities.maxImageCount;
-    }
-
-    VkSwapchainCreateInfoKHR swapchainInfo{};
-    swapchainInfo.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
-    swapchainInfo.surface = platformSurface_;
-    swapchainInfo.minImageCount = imageCount;
-    swapchainInfo.imageFormat = surfaceFormat.format;
-    swapchainInfo.imageColorSpace = surfaceFormat.colorSpace;
-    swapchainInfo.imageExtent = extent;
-    swapchainInfo.imageArrayLayers = 1;
-    swapchainInfo.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-
-    uint32_t queueFamilyIndices[] = { graphicsQueueFamily_, presentQueueFamily_ };
-    if (graphicsQueueFamily_ != presentQueueFamily_) {
-        swapchainInfo.imageSharingMode = VK_SHARING_MODE_CONCURRENT;
-        swapchainInfo.queueFamilyIndexCount = 2;
-        swapchainInfo.pQueueFamilyIndices = queueFamilyIndices;
-    } else {
-        swapchainInfo.imageSharingMode = VK_SHARING_MODE_EXCLUSIVE;
-        swapchainInfo.queueFamilyIndexCount = 0;
-        swapchainInfo.pQueueFamilyIndices = nullptr;
-    }
-
-    swapchainInfo.preTransform = support.capabilities.currentTransform;
-    swapchainInfo.compositeAlpha = VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR;
-    swapchainInfo.presentMode = presentMode;
-    swapchainInfo.clipped = VK_TRUE;
-    swapchainInfo.oldSwapchain = oldSwapchain;
-
-    VkResult result = vkCreateSwapchainKHR(device_, &swapchainInfo, nullptr, &swapchain_);
-    if (result != VK_SUCCESS || swapchain_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create swapchain (VkResult %d).\n", static_cast<int>(result));
-        swapchain_ = VK_NULL_HANDLE;
-        return false;
-    }
-
-    vkGetSwapchainImagesKHR(device_, swapchain_, &imageCount, nullptr);
-    swapchainImages_.resize(imageCount);
-    vkGetSwapchainImagesKHR(device_, swapchain_, &imageCount, swapchainImages_.data());
-
-    swapchainFormat_ = surfaceFormat.format;
-    swapchainExtent_ = extent;
-
-    swapchainImageViews_.resize(imageCount);
-    for (uint32_t i = 0; i < imageCount; ++i) {
-        VkImageViewCreateInfo viewInfo{};
-        viewInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-        viewInfo.image = swapchainImages_[i];
-        viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-        viewInfo.format = swapchainFormat_;
-        viewInfo.components.r = VK_COMPONENT_SWIZZLE_IDENTITY;
-        viewInfo.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-        viewInfo.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-        viewInfo.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-        viewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-        viewInfo.subresourceRange.baseMipLevel = 0;
-        viewInfo.subresourceRange.levelCount = 1;
-        viewInfo.subresourceRange.baseArrayLayer = 0;
-        viewInfo.subresourceRange.layerCount = 1;
-
-        VkResult viewResult = vkCreateImageView(device_, &viewInfo, nullptr, &swapchainImageViews_[i]);
-        if (viewResult != VK_SUCCESS || swapchainImageViews_[i] == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create swapchain image view (VkResult %d).\n", static_cast<int>(viewResult));
-            return false;
-        }
-    }
-
-    VkAttachmentDescription colorAttachment{};
-    colorAttachment.format = swapchainFormat_;
-    colorAttachment.samples = VK_SAMPLE_COUNT_1_BIT;
-    colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    colorAttachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    colorAttachment.finalLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-
-    VkAttachmentReference colorAttachmentRef{};
-    colorAttachmentRef.attachment = 0;
-    colorAttachmentRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkSubpassDescription subpass{};
-    subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    subpass.colorAttachmentCount = 1;
-    subpass.pColorAttachments = &colorAttachmentRef;
-
-    VkSubpassDependency dependency{};
-    dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-    dependency.dstSubpass = 0;
-    dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-    dependency.srcAccessMask = 0;
-    dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-    dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-
-    VkRenderPassCreateInfo renderPassInfo{};
-    renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-    renderPassInfo.attachmentCount = 1;
-    renderPassInfo.pAttachments = &colorAttachment;
-    renderPassInfo.subpassCount = 1;
-    renderPassInfo.pSubpasses = &subpass;
-    renderPassInfo.dependencyCount = 1;
-    renderPassInfo.pDependencies = &dependency;
-
-    VkResult renderPassResult = vkCreateRenderPass(device_, &renderPassInfo, nullptr, &renderPass_);
-    if (renderPassResult != VK_SUCCESS || renderPass_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create render pass (VkResult %d).\n", static_cast<int>(renderPassResult));
-        return false;
-    }
-
-    swapchainFramebuffers_.resize(imageCount);
-    for (uint32_t i = 0; i < imageCount; ++i) {
-        VkImageView attachments[] = { swapchainImageViews_[i] };
-
-        VkFramebufferCreateInfo framebufferInfo{};
-        framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
-        framebufferInfo.renderPass = renderPass_;
-        framebufferInfo.attachmentCount = 1;
-        framebufferInfo.pAttachments = attachments;
-        framebufferInfo.width = swapchainExtent_.width;
-        framebufferInfo.height = swapchainExtent_.height;
-        framebufferInfo.layers = 1;
-
-        VkResult framebufferResult = vkCreateFramebuffer(device_, &framebufferInfo, nullptr, &swapchainFramebuffers_[i]);
-        if (framebufferResult != VK_SUCCESS || swapchainFramebuffers_[i] == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create framebuffer (VkResult %d).\n", static_cast<int>(framebufferResult));
-            return false;
-        }
-    }
-
-    imagesInFlight_.assign(imageCount, VK_NULL_HANDLE);
-    vsyncEnabled_ = (presentMode == VK_PRESENT_MODE_FIFO_KHR || presentMode == VK_PRESENT_MODE_FIFO_RELAXED_KHR);
-    if (vsyncEnabled_) {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags | QVF_VIDEOSYNC);
-    } else {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags & ~QVF_VIDEOSYNC);
-    }
-
-    return true;
-}
-
-bool VulkanRenderer::createDescriptorPool() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    if (descriptorPool_ != VK_NULL_HANDLE) {
-        return true;
-    }
-
-    std::array<VkDescriptorPoolSize, 3> poolSizes{};
-    poolSizes[0].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    poolSizes[0].descriptorCount = 128;
-    poolSizes[1].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    poolSizes[1].descriptorCount = 256;
-    poolSizes[2].type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    poolSizes[2].descriptorCount = 64;
-
-    VkDescriptorPoolCreateInfo poolInfo{};
-    poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    poolInfo.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-    poolInfo.maxSets = 512;
-    poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
-    poolInfo.pPoolSizes = poolSizes.data();
-
-    VkResult result = vkCreateDescriptorPool(device_, &poolInfo, nullptr, &descriptorPool_);
-    if (result != VK_SUCCESS || descriptorPool_ == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: failed to create descriptor pool (VkResult %d).\n", static_cast<int>(result));
-        descriptorPool_ = VK_NULL_HANDLE;
-        return false;
-    }
-
-    return true;
-}
-
-bool VulkanRenderer::createSyncObjects() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    inFlightFrames_.resize(kMaxFramesInFlight);
-
-    VkSemaphoreCreateInfo semaphoreInfo{};
-    semaphoreInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-
-    VkFenceCreateInfo fenceInfo{};
-    fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-    fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-
-    for (InFlightFrame &frame : inFlightFrames_) {
-        VkResult semaphoreResult = vkCreateSemaphore(device_, &semaphoreInfo, nullptr, &frame.imageAvailable);
-        if (semaphoreResult != VK_SUCCESS || frame.imageAvailable == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create image-available semaphore (VkResult %d).\n", static_cast<int>(semaphoreResult));
-            return false;
-        }
-
-        VkResult renderSemaphoreResult = vkCreateSemaphore(device_, &semaphoreInfo, nullptr, &frame.renderFinished);
-        if (renderSemaphoreResult != VK_SUCCESS || frame.renderFinished == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create render-finished semaphore (VkResult %d).\n", static_cast<int>(renderSemaphoreResult));
-            return false;
-        }
-
-        VkResult fenceResult = vkCreateFence(device_, &fenceInfo, nullptr, &frame.inFlight);
-        if (fenceResult != VK_SUCCESS || frame.inFlight == VK_NULL_HANDLE) {
-            Com_Printf("refresh-vk: failed to create in-flight fence (VkResult %d).\n", static_cast<int>(fenceResult));
-            return false;
-        }
-
-        frame.commandBuffer = VK_NULL_HANDLE;
-        frame.imageIndex = 0;
-        frame.hasImage = false;
-    }
-
-    if (!inFlightFrames_.empty()) {
-        std::vector<VkCommandBuffer> buffers(inFlightFrames_.size(), VK_NULL_HANDLE);
-        VkCommandBufferAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-        allocInfo.commandPool = commandPool_;
-        allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-        allocInfo.commandBufferCount = static_cast<uint32_t>(buffers.size());
-
-        VkResult allocResult = vkAllocateCommandBuffers(device_, &allocInfo, buffers.data());
-        if (allocResult != VK_SUCCESS) {
-            Com_Printf("refresh-vk: failed to allocate command buffers (VkResult %d).\n", static_cast<int>(allocResult));
-            return false;
-        }
-
-        for (size_t i = 0; i < inFlightFrames_.size(); ++i) {
-            inFlightFrames_[i].commandBuffer = buffers[i];
-        }
-    }
-
-    currentFrameIndex_ = 0;
-    return true;
-}
-
-bool VulkanRenderer::createDeviceResources() {
-    if (!pickPhysicalDevice()) {
-        return false;
-    }
-
-    if (!createLogicalDevice()) {
-        return false;
-    }
-
-    if (!createCommandPool()) {
-        return false;
-    }
-
-    if (!createSwapchainResources()) {
-        return false;
-    }
-
-    if (!createDescriptorPool()) {
-        return false;
-    }
-
-    if (!createTextureDescriptorSetLayout()) {
-        destroyDescriptorPool();
-    }
-  
-    if (!createModelDescriptorResources()) {
-        return false;
-    }
-
-    if (!createSyncObjects()) {
-        destroySwapchainResources();
-        return false;
-    }
-
-    return true;
-}
-
-void VulkanRenderer::destroySyncObjects() {
-    if (device_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    for (InFlightFrame &frame : inFlightFrames_) {
-        if (frame.commandBuffer != VK_NULL_HANDLE && commandPool_ != VK_NULL_HANDLE) {
-            vkFreeCommandBuffers(device_, commandPool_, 1, &frame.commandBuffer);
-            frame.commandBuffer = VK_NULL_HANDLE;
-        } else {
-            frame.commandBuffer = VK_NULL_HANDLE;
-        }
-        if (frame.imageAvailable != VK_NULL_HANDLE) {
-            vkDestroySemaphore(device_, frame.imageAvailable, nullptr);
-            frame.imageAvailable = VK_NULL_HANDLE;
-        }
-        if (frame.renderFinished != VK_NULL_HANDLE) {
-            vkDestroySemaphore(device_, frame.renderFinished, nullptr);
-            frame.renderFinished = VK_NULL_HANDLE;
-        }
-        if (frame.inFlight != VK_NULL_HANDLE) {
-            vkDestroyFence(device_, frame.inFlight, nullptr);
-            frame.inFlight = VK_NULL_HANDLE;
-        }
-        frame.hasImage = false;
-    }
-
-    inFlightFrames_.clear();
-    lastSubmittedFrame_.reset();
-}
-
-void VulkanRenderer::destroySwapchainResources() {
-    if (device_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    destroySyncObjects();
-
-    for (VkFramebuffer framebuffer : swapchainFramebuffers_) {
-        if (framebuffer != VK_NULL_HANDLE) {
-            vkDestroyFramebuffer(device_, framebuffer, nullptr);
-        }
-    }
-    swapchainFramebuffers_.clear();
-
-    if (renderPass_ != VK_NULL_HANDLE) {
-        vkDestroyRenderPass(device_, renderPass_, nullptr);
-        renderPass_ = VK_NULL_HANDLE;
-    }
-
-    for (VkImageView view : swapchainImageViews_) {
-        if (view != VK_NULL_HANDLE) {
-            vkDestroyImageView(device_, view, nullptr);
-        }
-    }
-    swapchainImageViews_.clear();
-
-    if (swapchain_ != VK_NULL_HANDLE) {
-        vkDestroySwapchainKHR(device_, swapchain_, nullptr);
-        swapchain_ = VK_NULL_HANDLE;
-    }
-
-    swapchainImages_.clear();
-    imagesInFlight_.clear();
-    frameAcquired_ = false;
-    frameActive_ = false;
-    swapchainExtent_ = { 0u, 0u };
-    swapchainFormat_ = VK_FORMAT_UNDEFINED;
-    vsyncEnabled_ = (swapInterval_ > 0);
-    if (vsyncEnabled_) {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags | QVF_VIDEOSYNC);
-    } else {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags & ~QVF_VIDEOSYNC);
-    }
-    lastSubmittedFrame_.reset();
-}
-
-void VulkanRenderer::destroyDescriptorPool() {
-    if (descriptorPool_ != VK_NULL_HANDLE && device_ != VK_NULL_HANDLE) {
-        vkDestroyDescriptorPool(device_, descriptorPool_, nullptr);
-        descriptorPool_ = VK_NULL_HANDLE;
-    }
-}
-
-void VulkanRenderer::destroyCommandPool() {
-    if (commandPool_ != VK_NULL_HANDLE && device_ != VK_NULL_HANDLE) {
-        vkDestroyCommandPool(device_, commandPool_, nullptr);
-        commandPool_ = VK_NULL_HANDLE;
-    }
-}
-
-void VulkanRenderer::destroyDeviceResources() {
-    if (device_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    vkDeviceWaitIdle(device_);
-    clear2DBatches();
-    destroyAllModelGeometry();
-    destroySwapchainResources();
-    destroyAllImageResources();
-    destroyModelDescriptorResources();
-    destroyDescriptorPool();
-    destroyTextureDescriptorSetLayout();
-    destroyCommandPool();
-
-    if (device_ != VK_NULL_HANDLE) {
-        vkDestroyDevice(device_, nullptr);
-        device_ = VK_NULL_HANDLE;
-    }
-
-    graphicsQueue_ = VK_NULL_HANDLE;
-    presentQueue_ = VK_NULL_HANDLE;
-    graphicsQueueFamily_ = VK_QUEUE_FAMILY_IGNORED;
-    presentQueueFamily_ = VK_QUEUE_FAMILY_IGNORED;
-    physicalDevice_ = VK_NULL_HANDLE;
-    memoryProperties_ = {};
-    physicalDeviceProperties_ = {};
-    supportedFeatures_ = {};
-    enabledFeatures_ = {};
-    deviceExtensions_.clear();
-    lastSubmittedFrame_.reset();
-}
-
-bool VulkanRenderer::rebuildSwapchain() {
-    if (device_ == VK_NULL_HANDLE) {
-        return false;
-    }
-
-    vkDeviceWaitIdle(device_);
-
-    destroySwapchainResources();
-
-    if (!createSwapchainResources()) {
-        return false;
-    }
-
-    if (!createSyncObjects()) {
-        destroySwapchainResources();
-        return false;
-    }
-
-    return true;
-}
-
-bool VulkanRenderer::recreateSwapchain() {
-    return rebuildSwapchain();
-}
-
-void VulkanRenderer::destroyVulkan() {
-    destroyDeviceResources();
-    destroyPlatformSurface(instance_, nullptr);
-    destroyInstance();
-}
-
-void VulkanRenderer::finishFrameRecording() {
-    if (!frameActive_ || !frameAcquired_) {
-        return;
-    }
-
-    InFlightFrame &frame = inFlightFrames_[currentFrameIndex_];
-    if (!frame.hasImage || frame.commandBuffer == VK_NULL_HANDLE) {
-        return;
-    }
-
-    vkCmdEndRenderPass(frame.commandBuffer);
-    VkResult endResult = vkEndCommandBuffer(frame.commandBuffer);
-    if (endResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to end command buffer (VkResult %d).\n", static_cast<int>(endResult));
-    }
-}
-
-bool VulkanRenderer::init(bool total) {
-    if (!total) {
-        frameActive_ = false;
-        return true;
-    }
-
-    if (initialized_) {
-        return true;
-    }
-
-    Renderer_InitSharedCvars();
-
-    vk_fog = Cvar_Get("vk_fog", "-1", 0);
-    vk_bloom = Cvar_Get("vk_bloom", "-1", 0);
-    vk_polyblend = Cvar_Get("vk_polyblend", "-1", 0);
-    vk_waterwarp = Cvar_Get("vk_waterwarp", "-1", 0);
-    vk_dynamic = Cvar_Get("vk_dynamic", "-1", 0);
-    vk_perPixelLighting = Cvar_Get("vk_per_pixel_lighting", "-1", 0);
-
-    Com_Printf("------- refresh-vk init -------\n");
-
-    initializePlatformHooks();
-    collectPlatformInstanceExtensions();
-
-    VideoGeometry geometry = queryVideoGeometry();
-    applyVideoGeometry(geometry);
-    Com_Printf("Using video geometry %dx%d%s\n", r_config.width, r_config.height,
-               (r_config.flags & QVF_FULLSCREEN) ? " (fullscreen)" : "");
-
-    resetTransientState();
-
-    swapInterval_ = readSwapIntervalSetting();
-    if (swapInterval_ > 0) {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags | QVF_VIDEOSYNC);
-    } else {
-        r_config.flags = static_cast<vidFlags_t>(r_config.flags & ~QVF_VIDEOSYNC);
-    }
-
-    updateUIScaling();
-
-    models_.clear();
-    images_.clear();
-    modelLookup_.clear();
-    imageLookup_.clear();
-    rawPic_ = {};
-
-    if (!draw2d::initialize()) {
-        Com_Printf("Failed to initialize Vulkan 2D helper.\n");
-        return false;
-    }
-
-    if (!createInstance()) {
-        draw2d::shutdown();
-        Com_Printf("refresh-vk: unable to create Vulkan instance.\n");
-        return false;
-    }
-
-    if (!createPlatformSurface(instance_, nullptr)) {
-        destroyVulkan();
-        draw2d::shutdown();
-        return false;
-    }
-
-    if (!createDeviceResources()) {
-        destroyVulkan();
-        draw2d::shutdown();
-        return false;
-    }
-
-    if (!platformInstanceExtensions_.empty()) {
-        Com_DPrintf("refresh-vk: required platform instance extensions:\n");
-        for (const std::string &extension : platformInstanceExtensions_) {
-            Com_DPrintf("    %s\n", extension.c_str());
-        }
-    }
-
-    initialized_ = true;
-    r_registration_sequence = 1;
-
-    Com_Printf("refresh-vk initialized.\n");
-    Com_Printf("------------------------------\n");
-
-    return true;
-}
-
-void VulkanRenderer::shutdown(bool total) {
-    if (!initialized_) {
-        destroyVulkan();
-        draw2d::shutdown();
-        platformInstanceExtensions_.clear();
-        platformHooks_ = {};
-        platformInstance_ = VK_NULL_HANDLE;
-        return;
-    }
-
-    if (frameActive_) {
-        draw2d::end();
-        finishFrameRecording();
-        frameActive_ = false;
-    }
-
-    frameAcquired_ = false;
-
-    destroyVulkan();
-
-    platformInstanceExtensions_.clear();
-    platformHooks_ = {};
-    platformInstance_ = VK_NULL_HANDLE;
-
-    models_.clear();
-    images_.clear();
-    modelLookup_.clear();
-    imageLookup_.clear();
-    rawPic_ = {};
-    whiteTextureHandle_ = 0;
-    rawTextureHandle_ = 0;
-    currentMap_.clear();
-    resetTransientState();
-    draw2d::shutdown();
-
-    if (total) {
-        initialized_ = false;
-    } else {
-        initialized_ = false;
-    }
-}
-
-void VulkanRenderer::beginRegistration(const char *map) {
-    if (map) {
-        currentMap_ = map;
-    } else {
-        currentMap_.clear();
-    }
-
-    ++r_registration_sequence;
-}
-
-qhandle_t VulkanRenderer::registerModel(const char *name) {
-    if (!name || !*name) {
-        return 0;
-    }
-
-    if (name[0] == '*') {
-        int inlineIndex = Q_atoi(name + 1);
-        if (inlineIndex < 0) {
-            return 0;
-        }
-
-        qhandle_t handle = ~inlineIndex;
-        std::string key{name};
-        auto [it, inserted] = modelLookup_.emplace(key, handle);
-        if (!inserted) {
-            handle = it->second;
-        }
-
-        auto &record = models_[handle];
-        record.handle = handle;
-        record.name = name;
-        record.registrationSequence = r_registration_sequence;
-        record.type = 0;
-        record.numFrames = 0;
-        record.numMeshes = 0;
-        record.inlineModel = true;
-        record.aliasFrames.clear();
-        record.spriteFrames.clear();
-        destroyModelRecord(record);
-
-        return handle;
-    }
-
-    qhandle_t handle = registerResource(modelLookup_, name);
-    auto &record = models_[handle];
-    record.handle = handle;
-    record.name = name;
-    record.registrationSequence = r_registration_sequence;
-    record.inlineModel = false;
-    record.aliasFrames.clear();
-    record.spriteFrames.clear();
-    destroyModelRecord(record);
-    record.type = 0;
-    record.numFrames = 0;
-    record.numMeshes = 0;
-
-    model_t *model = nullptr;
-    char normalized[MAX_QPATH];
-    size_t normalizedLength = FS_NormalizePathBuffer(normalized, name, MAX_QPATH);
-    if (normalizedLength > 0 && normalizedLength < MAX_QPATH) {
-        model = MOD_Find(normalized);
-    }
-    if (!model) {
-        model = MOD_Find(name);
-    }
-    if (!model && handle > 0) {
-        model_t *fromHandle = MOD_ForHandle(handle);
-        if (fromHandle) {
-            if (normalizedLength > 0 && normalizedLength < MAX_QPATH) {
-                if (FS_pathcmp(fromHandle->name, normalized) == 0) {
-                    model = fromHandle;
-                }
-            } else {
-                model = fromHandle;
-            }
-        }
-    }
-
-    if (model) {
-        record.type = model->type;
-        record.numFrames = model->numframes;
-        record.numMeshes = model->nummeshes;
-
-        if (model->type == MOD_ALIAS && model->frames) {
-            record.aliasFrames.reserve(static_cast<size_t>(model->numframes));
-            for (int i = 0; i < model->numframes; ++i) {
-                record.aliasFrames.emplace_back(makeAliasFrameMetadata(model->frames[i]));
-            }
-        } else if (model->type == MOD_SPRITE && model->spriteframes) {
-            record.spriteFrames.reserve(static_cast<size_t>(model->numframes));
-            for (int i = 0; i < model->numframes; ++i) {
-                const mspriteframe_t &frame = model->spriteframes[i];
-                ModelRecord::SpriteFrameMetadata metadata{};
-                metadata.width = frame.width;
-                metadata.height = frame.height;
-                metadata.originX = frame.origin_x;
-                metadata.originY = frame.origin_y;
-                record.spriteFrames.emplace_back(metadata);
-            }
-        }
-
-        allocateModelGeometry(record, *model);
-    }
-
-    return handle;
-}
-
-void VulkanRenderer::allocateModelGeometry(ModelRecord &record, const model_t &model) {
-    destroyModelRecord(record);
-
-    if (model.type != MOD_ALIAS) {
-        return;
-    }
-
-    if (!model.meshes || model.nummeshes <= 0) {
-        return;
-    }
-
-    record.meshGeometry.reserve(static_cast<size_t>(model.nummeshes));
-
-    for (int meshIndex = 0; meshIndex < model.nummeshes; ++meshIndex) {
-        const maliasmesh_t &mesh = model.meshes[meshIndex];
-        ModelRecord::MeshGeometry geometry{};
-        geometry.descriptor.binding = static_cast<uint32_t>(meshIndex);
-
-        if (mesh.numverts > 0 && mesh.verts && mesh.tcoords) {
-            const size_t vertexSize = static_cast<size_t>(mesh.numverts) * sizeof(*mesh.verts);
-            const size_t texCoordSize = static_cast<size_t>(mesh.numverts) * sizeof(*mesh.tcoords);
-            geometry.vertexStaging.resize(vertexSize + texCoordSize);
-            std::memcpy(geometry.vertexStaging.data(), mesh.verts, vertexSize);
-            std::memcpy(geometry.vertexStaging.data() + vertexSize, mesh.tcoords, texCoordSize);
-            geometry.vertexCount = static_cast<size_t>(mesh.numverts);
-        }
-
-        if (mesh.numindices > 0 && mesh.indices) {
-            const size_t indexSize = static_cast<size_t>(mesh.numindices) * sizeof(*mesh.indices);
-            geometry.indexStaging.resize(indexSize);
-            std::memcpy(geometry.indexStaging.data(), mesh.indices, indexSize);
-            geometry.indexCount = static_cast<size_t>(mesh.numindices);
-            geometry.indexType = VK_INDEX_TYPE_UINT16;
-        }
-
-        if (geometry.vertexCount == 0 && geometry.indexCount == 0) {
-            continue;
-        }
-
-        if (!geometry.vertexStaging.empty() || !geometry.indexStaging.empty()) {
-            if (device_ != VK_NULL_HANDLE && !uploadMeshGeometry(geometry)) {
-                Com_Printf("refresh-vk: failed to upload mesh %d for model %s.\n",
-                           meshIndex,
-                           record.name.c_str());
-            }
-        }
-
-        record.meshGeometry.emplace_back(std::move(geometry));
-    }
-}
-
-void VulkanRenderer::bindModelGeometryBuffers(ModelRecord &record) {
-    if (record.meshGeometry.empty()) {
-        return;
-    }
-
-    if (inFlightFrames_.empty() || !frameActive_) {
-        return;
-    }
-
-    InFlightFrame &frame = inFlightFrames_[currentFrameIndex_];
-    VkCommandBuffer commandBuffer = frame.commandBuffer;
-    if (commandBuffer == VK_NULL_HANDLE) {
-        return;
-    }
-
-    for (size_t meshIndex = 0; meshIndex < record.meshGeometry.size(); ++meshIndex) {
-        auto &geometry = record.meshGeometry[meshIndex];
-
-        if (!geometry.uploaded && (!geometry.vertexStaging.empty() || !geometry.indexStaging.empty())) {
-            if (!uploadMeshGeometry(geometry)) {
-                continue;
-            }
-        }
-
-        if (geometry.vertex.buffer == VK_NULL_HANDLE && geometry.index.buffer == VK_NULL_HANDLE) {
-            continue;
-        }
-
-        std::string entry{"bind.model."};
-        entry.append(record.name);
-        entry.push_back('#');
-        entry.append(std::to_string(meshIndex));
-        commandLog_.push_back(std::move(entry));
-
-        if (geometry.vertex.buffer != VK_NULL_HANDLE && geometry.vertex.size > 0) {
-            VkBuffer buffers[] = { geometry.vertex.buffer };
-            VkDeviceSize offsets[] = { geometry.vertex.offset };
-            vkCmdBindVertexBuffers(commandBuffer, 0, 1, buffers, offsets);
-        }
-
-        if (geometry.index.buffer != VK_NULL_HANDLE && geometry.index.size > 0) {
-            vkCmdBindIndexBuffer(commandBuffer, geometry.index.buffer, geometry.index.offset, geometry.indexType);
-        }
-
-        if (geometry.descriptor.set != VK_NULL_HANDLE && modelPipelineLayout_ != VK_NULL_HANDLE) {
-            vkCmdBindDescriptorSets(commandBuffer,
-                                    VK_PIPELINE_BIND_POINT_GRAPHICS,
-                                    modelPipelineLayout_,
-                                    0,
-                                    1,
-                                    &geometry.descriptor.set,
-                                    0,
-                                    nullptr);
-        }
-    }
-}
-
-qhandle_t VulkanRenderer::registerImage(const char *name, imagetype_t type, imageflags_t flags) {
-    if (!name || !*name) {
-        return 0;
-    }
-
-    image_pixels_t pixels{};
-    if (!IMG_LoadPixels(name, type, flags, &pixels)) {
-        return 0;
-    }
-
-    std::string canonical = pixels.name[0] ? pixels.name : std::string{name};
-    qhandle_t handle = registerResource(imageLookup_, canonical);
-    if (!handle) {
-        IMG_FreePixels(&pixels);
-        return 0;
-    }
-
-    auto &record = images_[handle];
-    destroyImageRecord(record);
-
-    record.handle = handle;
-    record.name = canonical;
-    record.type = static_cast<imagetype_t>(pixels.type);
-    record.flags = static_cast<imageflags_t>(pixels.flags);
-    record.width = pixels.width;
-    record.height = pixels.height;
-    record.uploadWidth = pixels.upload_width;
-    record.uploadHeight = pixels.upload_height;
-    record.transparent = (record.flags & IF_TRANSPARENT) != 0;
-    record.registrationSequence = r_registration_sequence;
-
-    uint32_t texWidth = static_cast<uint32_t>(std::max(1, record.uploadWidth));
-    uint32_t texHeight = static_cast<uint32_t>(std::max(1, record.uploadHeight));
-    VkFormat format = VK_FORMAT_R8G8B8A8_UNORM;
-    const uint8_t *pixelData = reinterpret_cast<const uint8_t *>(pixels.pixels);
-    size_t pixelSize = pixels.size;
-
-    bool uploaded = (pixelData && pixelSize > 0) &&
-                    ensureTextureResources(record, pixelData, pixelSize, texWidth, texHeight, format);
-
-    IMG_FreePixels(&pixels);
-
-    if (!uploaded) {
-        images_.erase(handle);
-        imageLookup_.erase(canonical);
-        imageLookup_.erase(std::string{name});
-        return 0;
-    }
-
-    imageLookup_[canonical] = handle;
-    if (canonical != name) {
-        imageLookup_[std::string{name}] = handle;
-    }
-
-    return handle;
-}
-
-void VulkanRenderer::setSky(const char *name, float rotate, bool autorotate, const vec3_t axis) {
-    sky_.name = name ? name : "";
-    sky_.rotate = rotate;
-    sky_.autorotate = autorotate;
-    if (axis) {
-        std::copy_n(axis, sky_.axis.size(), sky_.axis.begin());
-    }
-}
-
-void VulkanRenderer::endRegistration() {
-    for (auto it = models_.begin(); it != models_.end();) {
-        if (it->second.registrationSequence != r_registration_sequence) {
-            destroyModelRecord(it->second);
-            it = models_.erase(it);
-        } else {
-            ++it;
-        }
-    }
-
-    auto predicate = [](auto &pair) {
-        return pair.second.registrationSequence != r_registration_sequence;
-    };
-
-    std::erase_if(images_, predicate);
-}
-
-void VulkanRenderer::beginFrame() {
-    if (!initialized_ || device_ == VK_NULL_HANDLE || swapchain_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    refreshSwapInterval();
-    if (swapchain_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    if (vid && vid->pump_events) {
-        vid->pump_events();
-    }
-
-    resetFrameState();
-    commandLog_.clear();
-
-    if (inFlightFrames_.empty()) {
-        return;
-    }
-
-    InFlightFrame &frame = inFlightFrames_[currentFrameIndex_];
-    if (frame.inFlight != VK_NULL_HANDLE) {
-        vkWaitForFences(device_, 1, &frame.inFlight, VK_TRUE, std::numeric_limits<uint64_t>::max());
-    }
-
-    if (lastSubmittedFrame_.has_value() && *lastSubmittedFrame_ == currentFrameIndex_) {
-        lastSubmittedFrame_.reset();
-    }
-
-    clear2DBatches();
-
-    uint32_t imageIndex = 0;
-    VkResult acquireResult = vkAcquireNextImageKHR(device_, swapchain_, std::numeric_limits<uint64_t>::max(), frame.imageAvailable, VK_NULL_HANDLE, &imageIndex);
-
-    if (acquireResult == VK_ERROR_OUT_OF_DATE_KHR) {
-        rebuildSwapchain();
-        return;
-    }
-
-    if (acquireResult != VK_SUCCESS && acquireResult != VK_SUBOPTIMAL_KHR) {
-        Com_Printf("refresh-vk: failed to acquire swapchain image (VkResult %d).\n", static_cast<int>(acquireResult));
-        return;
-    }
-
-    frame.imageIndex = imageIndex;
-    frame.hasImage = true;
-
-    if (imageIndex < imagesInFlight_.size() && imagesInFlight_[imageIndex] != VK_NULL_HANDLE) {
-        vkWaitForFences(device_, 1, &imagesInFlight_[imageIndex], VK_TRUE, std::numeric_limits<uint64_t>::max());
-    }
-
-    if (imageIndex < imagesInFlight_.size()) {
-        imagesInFlight_[imageIndex] = frame.inFlight;
-    }
-
-    if (frame.inFlight != VK_NULL_HANDLE) {
-        vkResetFences(device_, 1, &frame.inFlight);
-    }
-
-    if (frame.commandBuffer == VK_NULL_HANDLE) {
-        return;
-    }
-
-    VkResult resetResult = vkResetCommandBuffer(frame.commandBuffer, 0);
-    if (resetResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to reset command buffer (VkResult %d).\n", static_cast<int>(resetResult));
-        return;
-    }
-
-    VkCommandBufferBeginInfo beginInfo{};
-    beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-    beginInfo.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-
-    VkResult beginResult = vkBeginCommandBuffer(frame.commandBuffer, &beginInfo);
-    if (beginResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to begin command buffer (VkResult %d).\n", static_cast<int>(beginResult));
-        return;
-    }
-
-    VkClearValue clearValue{};
-    clearValue.color = { { 0.0f, 0.0f, 0.0f, 1.0f } };
-
-    VkRenderPassBeginInfo renderPassInfo{};
-    renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
-    renderPassInfo.renderPass = renderPass_;
-    renderPassInfo.framebuffer = (frame.imageIndex < swapchainFramebuffers_.size()) ? swapchainFramebuffers_[frame.imageIndex] : VK_NULL_HANDLE;
-    renderPassInfo.renderArea.offset = { 0, 0 };
-    renderPassInfo.renderArea.extent = swapchainExtent_;
-    renderPassInfo.clearValueCount = 1;
-    renderPassInfo.pClearValues = &clearValue;
-
-    if (renderPassInfo.framebuffer == VK_NULL_HANDLE) {
-        Com_Printf("refresh-vk: framebuffer unavailable for frame.\n");
-        vkEndCommandBuffer(frame.commandBuffer);
-        return;
-    }
-
-    vkCmdBeginRenderPass(frame.commandBuffer, &renderPassInfo, VK_SUBPASS_CONTENTS_INLINE);
-
-    frameActive_ = true;
-    frameAcquired_ = true;
-
-    if (!draw2d::begin([this](const draw2d::Submission &submission) {
-            submit2DDraw(submission);
-        })) {
-        Com_Printf("vk_draw2d: failed to begin 2D batch for frame.\n");
-    }
-}
-
-void VulkanRenderer::endFrame() {
-    if (!initialized_ || !frameActive_) {
-        return;
-    }
-
-    draw2d::end();
-    finishFrameRecording();
-
-    if (inFlightFrames_.empty()) {
-        frameActive_ = false;
-        frameAcquired_ = false;
-        commandLog_.clear();
-        frameStats_.reset();
-        return;
-    }
-
-    InFlightFrame &frame = inFlightFrames_[currentFrameIndex_];
-    size_t submittedIndex = currentFrameIndex_;
-    if (!frame.hasImage || frame.commandBuffer == VK_NULL_HANDLE) {
-        frameActive_ = false;
-        frameAcquired_ = false;
-        commandLog_.clear();
-        frameStats_.reset();
-        return;
-    }
-
-    VkSemaphore waitSemaphores[] = { frame.imageAvailable };
-    VkPipelineStageFlags waitStages[] = { VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT };
-    VkSemaphore signalSemaphores[] = { frame.renderFinished };
-
-    VkSubmitInfo submitInfo{};
-    submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-    submitInfo.waitSemaphoreCount = 1;
-    submitInfo.pWaitSemaphores = waitSemaphores;
-    submitInfo.pWaitDstStageMask = waitStages;
-    submitInfo.commandBufferCount = 1;
-    submitInfo.pCommandBuffers = &frame.commandBuffer;
-    submitInfo.signalSemaphoreCount = 1;
-    submitInfo.pSignalSemaphores = signalSemaphores;
-
-    VkResult submitResult = vkQueueSubmit(graphicsQueue_, 1, &submitInfo, frame.inFlight);
-    if (submitResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to submit draw commands (VkResult %d).\n", static_cast<int>(submitResult));
-        if (frame.imageIndex < imagesInFlight_.size()) {
-            imagesInFlight_[frame.imageIndex] = VK_NULL_HANDLE;
-        }
-        vkQueueSubmit(graphicsQueue_, 0, nullptr, frame.inFlight);
-        vkWaitForFences(device_, 1, &frame.inFlight, VK_TRUE, std::numeric_limits<uint64_t>::max());
-        frame.hasImage = false;
-        frameActive_ = false;
-        frameAcquired_ = false;
-        commandLog_.clear();
-        frameStats_.reset();
-        return;
-    }
-
-    lastSubmittedFrame_ = submittedIndex;
-
-    VkSwapchainKHR swapchains[] = { swapchain_ };
-
-    VkPresentInfoKHR presentInfo{};
-    presentInfo.sType = VK_STRUCTURE_TYPE_PRESENT_INFO_KHR;
-    presentInfo.waitSemaphoreCount = 1;
-    presentInfo.pWaitSemaphores = signalSemaphores;
-    presentInfo.swapchainCount = 1;
-    presentInfo.pSwapchains = swapchains;
-    presentInfo.pImageIndices = &frame.imageIndex;
-
-    VkResult presentResult = vkQueuePresentKHR(presentQueue_, &presentInfo);
-    if (presentResult == VK_ERROR_OUT_OF_DATE_KHR || presentResult == VK_SUBOPTIMAL_KHR) {
-        recreateSwapchain();
-    } else if (presentResult != VK_SUCCESS) {
-        Com_Printf("refresh-vk: failed to present swapchain image (VkResult %d).\n", static_cast<int>(presentResult));
-    }
-
-    commandLog_.clear();
-    frameStats_.reset();
-
-    frame.hasImage = false;
-    frameActive_ = false;
-    frameAcquired_ = false;
-    currentFrameIndex_ = (currentFrameIndex_ + 1) % inFlightFrames_.size();
-}
-
-void VulkanRenderer::renderFrame(const refdef_t *fd) {
-    if (!initialized_ || !fd) {
-        return;
-    }
-
-    if (fd->height > 0) {
-        autoScaleValue_ = std::max(1, fd->height / SCREEN_HEIGHT);
-    }
-
-    prepareFrameState(*fd);
-    evaluateFrameSettings();
-    uploadDynamicLights();
-    updateSkyState();
-
-    beginWorldPass();
-    if (!(frameState_.refdef.rdflags & RDF_NOWORLDMODEL)) {
-        renderWorld();
-    }
-    endWorldPass();
-    recordStage("frame.begin");
-
-    classifyEntities(*fd);
-    sortTransparentQueues(frameState_.refdef);
-    buildEffectBuffers(*fd);
-
-    if (!(fd->rdflags & RDF_NOWORLDMODEL)) {
-        recordStage("world.draw");
-    }
-
-    auto processQueue = [&](const std::vector<const entity_t *> &queue, std::string_view label) {
-        if (queue.empty()) {
-            return;
-        }
-
-        PipelineKey currentKey{};
-        bool havePipeline = false;
-        size_t batchCount = 0;
-        const PipelineDesc *pipeline = nullptr;
-
-        for (auto it = queue.rbegin(); it != queue.rend(); ++it) {
-            const entity_t *entity = *it;
-            if (ModelRecord *record = findModelRecord(entity->model)) {
-                bindModelGeometryBuffers(*record);
-            }
-
-            PipelineKind kind = selectPipelineForEntity(*entity);
-            PipelineKey key = buildPipelineKey(kind);
-            if (!havePipeline || !(key == currentKey)) {
-                if (havePipeline && pipeline && batchCount) {
-                    recordDrawCall(*pipeline, label, batchCount);
-                }
-                currentKey = key;
-                pipeline = &ensurePipeline(currentKey);
-                havePipeline = true;
-                batchCount = 0;
-            }
-            ++batchCount;
-        }
-
-        if (havePipeline && pipeline && batchCount) {
-            recordDrawCall(*pipeline, label, batchCount);
-        }
-    };
-
-    processQueue(frameQueues_.bmodels, "entities.inline");
-    processQueue(frameQueues_.opaque, "entities.opaque");
-    processQueue(frameQueues_.alphaBack, "entities.alpha_back");
-
-    ViewParameters viewParams = computeViewParameters(*fd);
-
-    bool cylindricalBeams = gl_beamstyle && gl_beamstyle->integer != 0;
-    streamBeamPrimitives(viewParams, cylindricalBeams);
-    if (!effectStreams_.beamIndices.empty()) {
-        PipelineKind beamKind = cylindricalBeams ? PipelineKind::BeamCylindrical : PipelineKind::BeamSimple;
-        recordDrawCall(ensurePipeline(buildPipelineKey(beamKind)), "fx.beams", framePrimitives_.beams.size());
-    }
-
-    bool additiveParticles = gl_partstyle && gl_partstyle->integer != 0;
-    streamParticlePrimitives(viewParams, additiveParticles);
-    if (!effectStreams_.particleVertices.empty()) {
-        PipelineKind particleKind = additiveParticles ? PipelineKind::ParticleAdditive : PipelineKind::ParticleAlpha;
-        recordDrawCall(ensurePipeline(buildPipelineKey(particleKind)), "fx.particles", framePrimitives_.particles.size());
-    }
-
-    bool flaresEnabled = true;
-    if (cl_flares) {
-        flaresEnabled = cl_flares->integer != 0;
-    }
-    if (flaresEnabled) {
-        streamFlarePrimitives(viewParams);
-        if (!effectStreams_.flareVertices.empty()) {
-            recordDrawCall(ensurePipeline(buildPipelineKey(PipelineKind::Flare)), "fx.flares", framePrimitives_.flares.size());
-        }
-    } else {
-        effectStreams_.flareVertices.clear();
-        effectStreams_.flareIndices.clear();
-        frameStats_.flares = 0;
-    }
-
-    processQueue(frameQueues_.alphaFront, "entities.alpha_front");
-
-    streamDebugLinePrimitives();
-    if (!effectStreams_.debugLinesDepth.empty()) {
-        recordDrawCall(ensurePipeline(buildPipelineKey(PipelineKind::DebugLineDepth)), "debug.lines.depth", effectStreams_.debugLinesDepth.size() / 2);
-    }
-    if (!effectStreams_.debugLinesNoDepth.empty()) {
-        recordDrawCall(ensurePipeline(buildPipelineKey(PipelineKind::DebugLineNoDepth)), "debug.lines.nodepth", effectStreams_.debugLinesNoDepth.size() / 2);
-    }
-
-    bool waterwarpEnabled = resolveToggle(vk_waterwarp, "gl_waterwarp", true);
-    bool waterwarp = waterwarpEnabled && (fd->rdflags & RDF_UNDERWATER) != 0;
-    frameState_.waterwarpActive = waterwarp;
-    if (waterwarp) {
-        recordStage("post.waterwarp");
-    }
-
-    bool bloomEnabled = resolveToggle(vk_bloom, "gl_bloom", true);
-    bool worldVisible = !(fd->rdflags & RDF_NOWORLDMODEL) && frameState_.worldRendered;
-    bool bloom = bloomEnabled && worldVisible;
-    frameState_.bloomActive = bloom;
-    if (bloom) {
-        recordStage("post.bloom");
-    }
-
-    bool overlayEnabled = resolveToggle(vk_polyblend, "gl_polyblend", true);
-    bool overlayBlend = overlayEnabled && (fd->screen_blend[3] > 0.0f || fd->damage_blend[3] > 0.0f);
-    frameState_.overlayBlendActive = overlayBlend;
-    if (overlayBlend) {
-        recordStage("overlay.blend");
-    }
-
-    recordStage("frame.end");
-}
-
-void VulkanRenderer::lightPoint(const vec3_t origin, vec3_t light) const {
-    if (!light) {
-        return;
-    }
-
-    light[0] = 1.0f;
-    light[1] = 1.0f;
-    light[2] = 1.0f;
-
-    if (!frameState_.hasRefdef) {
-        return;
-    }
-
-    for (const auto &dlight : frameState_.dlights) {
-        float dx = origin[0] - dlight.origin[0];
-        float dy = origin[1] - dlight.origin[1];
-        float dz = origin[2] - dlight.origin[2];
-        float distance = std::sqrt(dx * dx + dy * dy + dz * dz);
-
-        float contribution = dlight.radius - DLIGHT_CUTOFF - distance;
-        if (contribution <= 0.0f) {
-            continue;
-        }
-
-        float scale = contribution * kInverseLightIntensity * dlight.intensity;
-        light[0] += scale * dlight.color[0];
-        light[1] += scale * dlight.color[1];
-        light[2] += scale * dlight.color[2];
-    }
-}
-
-VulkanRenderer::ScissorRect VulkanRenderer::fullScissorRect() const {
-    ScissorRect rect{};
-    rect.x = 0;
-    rect.y = 0;
-
-    auto computeExtent = [](int dimension, double scale) -> uint32_t {
-        if (dimension <= 0 || scale <= 0.0) {
-            return 0u;
-        }
-        double scaled = static_cast<double>(dimension) * scale;
-        long long rounded = std::llround(scaled);
-        if (rounded <= 0) {
-            return 0u;
-        }
-        return static_cast<uint32_t>(rounded);
-    };
-
-    double scale = static_cast<double>(std::max(scale_, 0.0f));
-    rect.width = computeExtent(r_config.width, scale);
-    rect.height = computeExtent(r_config.height, scale);
-    return rect;
-}
-
-std::optional<VulkanRenderer::ScissorRect> VulkanRenderer::scaledClipRect(const clipRect_t &clip) const {
-    ScissorRect bounds = fullScissorRect();
-    if (bounds.width == 0 || bounds.height == 0) {
-        return std::nullopt;
-    }
-
-    double scale = static_cast<double>(std::max(scale_, 0.0f));
-    if (scale <= 0.0) {
-        return std::nullopt;
-    }
-
-    double left = static_cast<double>(clip.left) * scale;
-    double top = static_cast<double>(clip.top) * scale;
-    double right = static_cast<double>(clip.right) * scale;
-    double bottom = static_cast<double>(clip.bottom) * scale;
-
-    int32_t x0 = static_cast<int32_t>(std::floor(left));
-    int32_t y0 = static_cast<int32_t>(std::floor(top));
-    int32_t x1 = static_cast<int32_t>(std::ceil(right));
-    int32_t y1 = static_cast<int32_t>(std::ceil(bottom));
-
-    int32_t maxWidth = static_cast<int32_t>(bounds.width);
-    int32_t maxHeight = static_cast<int32_t>(bounds.height);
-
-    x0 = std::clamp(x0, 0, maxWidth);
-    y0 = std::clamp(y0, 0, maxHeight);
-    x1 = std::clamp(x1, 0, maxWidth);
-    y1 = std::clamp(y1, 0, maxHeight);
-
-    if (x1 <= x0 || y1 <= y0) {
-        return std::nullopt;
-    }
-
-    ScissorRect rect{};
-    rect.x = x0;
-    rect.y = y0;
-    rect.width = static_cast<uint32_t>(x1 - x0);
-    rect.height = static_cast<uint32_t>(y1 - y0);
-    return rect;
-}
-
-void VulkanRenderer::recordScissorCommand(const ScissorRect &rect, bool clipped) {
-    std::string entry = "vkCmdSetScissor offset=(";
-    entry.append(std::to_string(rect.x));
-    entry.push_back(',');
-    entry.append(std::to_string(rect.y));
-    entry.append(") extent=(");
-    entry.append(std::to_string(rect.width));
-    entry.push_back(',');
-    entry.append(std::to_string(rect.height));
-    entry.push_back(')');
-    if (!clipped) {
-        entry.append(" (full)");
-    }
-    commandLog_.push_back(std::move(entry));
-}
-
-void VulkanRenderer::setClipRect(const clipRect_t *clip) {
-    bool changed = false;
-    if (clip) {
-        if (!clipRect_ || clipRect_->left != clip->left || clipRect_->right != clip->right ||
-            clipRect_->top != clip->top || clipRect_->bottom != clip->bottom) {
-            changed = true;
-        }
-    } else if (clipRect_) {
-        changed = true;
-    }
-
-    if (changed && draw2d::isActive()) {
-        draw2d::flush();
-    }
-
-    if (clip) {
-        clipRect_ = *clip;
-    } else {
-        clipRect_.reset();
-    }
-
-    ScissorRect fullRect = fullScissorRect();
-    std::optional<ScissorRect> desired = clip ? scaledClipRect(*clip) : std::nullopt;
-    bool drawingActive = frameActive_ && draw2d::isActive();
-
-    if (!desired) {
-        if (!activeScissor_) {
-            return;
-        }
-        if (drawingActive) {
-            draw2d::flush();
-            recordScissorCommand(fullRect, false);
-        }
-        activeScissor_.reset();
-        return;
-    }
-
-    if (activeScissor_ && *activeScissor_ == *desired) {
-        return;
-    }
-
-    if (drawingActive) {
-        draw2d::flush();
-        recordScissorCommand(*desired, true);
-    }
-
-    activeScissor_ = desired;
-}
-
-float VulkanRenderer::clampScale(cvar_t *var) const {
-    if (!var) {
-        return 1.0f;
-    }
-
-    float value = var->value;
-    if (!std::isfinite(value)) {
-        value = 1.0f;
-    }
-
-    constexpr float kMinScale = 0.25f;
-    constexpr float kMaxScale = 4.0f;
-    return std::clamp(value, kMinScale, kMaxScale);
-}
-
-void VulkanRenderer::setScale(float scale) {
-    float target = 1.0f;
-    if (std::isfinite(scale)) {
-        target = std::clamp(scale, 0.25f, 4.0f);
-    }
-
-    if (std::abs(target - scale_) <= std::numeric_limits<float>::epsilon()) {
-        return;
-    }
-
-    bool drawingActive = frameActive_ && draw2d::isActive();
-    if (drawingActive) {
-        draw2d::flush();
-    }
-
-    scale_ = target;
-    activeScissor_.reset();
-
-    if (clipRect_) {
-        setClipRect(&*clipRect_);
-    } else if (drawingActive) {
-        setClipRect(nullptr);
-    }
-}
-
-int VulkanRenderer::autoScale() const {
-    return autoScaleValue_;
-}
-
-void VulkanRenderer::drawChar(int x, int y, int flags, int ch, color_t color, qhandle_t font) {
-    drawStretchChar(x, y, kDefaultCharWidth, kDefaultCharHeight, flags, ch, color, font);
-}
-
-void VulkanRenderer::drawStretchChar(int x, int y, int w, int h, int flags, int ch, color_t color, qhandle_t font) {
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    if (!findImageRecord(font)) {
-        return;
-    }
-
-    uint8_t glyph = static_cast<uint8_t>(ch);
-    if ((glyph & 127u) == 32u) {
-        return;
-    }
-
-    if (flags & kUIAltColor) {
-        glyph |= 0x80u;
-    }
-    if (flags & kUIXorColor) {
-        glyph ^= 0x80u;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    float u0 = static_cast<float>(glyph & 15u) * kFontCellSize;
-    float v0 = static_cast<float>(glyph >> 4) * kFontCellSize;
-    auto uvs = makeUV(u0, v0, u0 + kFontCellSize, v0 + kFontCellSize);
-
-    if ((flags & kUIDropShadow) && glyph != 0x83u) {
-        float offset = kShadowOffset * scale;
-        color_t shadow = ColorA(color.a);
-        auto shadowPositions = makeQuad(fx + offset, fy + offset, fw, fh);
-        draw2d::submitQuad(shadowPositions, uvs, shadow.u32, font);
-    }
-
-    color_t finalColor = color;
-    if (glyph & 0x80u) {
-        finalColor = ColorSetAlpha(COLOR_WHITE, color.a);
-    }
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, uvs, finalColor.u32, font);
-}
-
-int VulkanRenderer::drawStringStretch(int x, int y, int scale, int flags, size_t maxChars,
-                                      const char *string, color_t color, qhandle_t font) {
-    if (!string || !*string) {
-        return x;
-    }
-
-    if (!canSubmit2D()) {
-        return x;
-    }
-
-    int effectiveScale = std::max(1, scale);
-    int charWidth = effectiveScale * kDefaultCharWidth;
-    int charHeight = effectiveScale * kDefaultCharHeight;
-    int cursorX = x;
-    int cursorY = y;
-    int lineStart = x;
-
-    size_t remaining = maxChars ? maxChars : std::numeric_limits<size_t>::max();
-    size_t processed = 0;
-    while (processed < remaining && string[processed]) {
-        unsigned char ch = static_cast<unsigned char>(string[processed]);
-        ++processed;
-
-        if ((flags & UI_MULTILINE) && ch == '\n') {
-            cursorY += charHeight;
-            cursorX = lineStart;
-            continue;
-        }
-
-        drawStretchChar(cursorX, cursorY, charWidth, charHeight, flags, ch, color, font);
-        cursorX += charWidth;
-    }
-
-    return cursorX;
-}
-
-int VulkanRenderer::drawKFontChar(int x, int y, int scale, int flags, uint32_t codepoint,
-                                  color_t color, const kfont_t *kfont) {
-    if (!kfont) {
-        return x;
-    }
-
-    const kfont_char_t *metrics = lookupKFontChar(kfont, codepoint);
-    if (!metrics) {
-        return x;
-    }
-
-    if (!canSubmit2D()) {
-        return x;
-    }
-
-    int effectiveScale = std::max(1, scale);
-    int w = static_cast<int>(metrics->w) * effectiveScale;
-    int h = static_cast<int>(metrics->h) * effectiveScale;
-    if (w <= 0 || h <= 0) {
-        return x;
-    }
-
-    const KFontRecord *record = findKFontRecord(kfont);
-    qhandle_t texture = record ? record->texture : kfont->pic;
-    if (!findImageRecord(texture)) {
-        return x;
-    }
-
-    float sw = record ? record->sw : kfont->sw;
-    float sh = record ? record->sh : kfont->sh;
-    float s0 = metrics->x * sw;
-    float t0 = metrics->y * sh;
-    float s1 = s0 + metrics->w * sw;
-    float t1 = t0 + metrics->h * sh;
-
-    float scaleFactor = scale_;
-    if (scaleFactor <= 0.0f) {
-        return x;
-    }
-
-    float fx = static_cast<float>(x) * scaleFactor;
-    float fy = static_cast<float>(y) * scaleFactor;
-    float fw = static_cast<float>(w) * scaleFactor;
-    float fh = static_cast<float>(h) * scaleFactor;
-
-    auto uvs = makeUV(s0, t0, s1, t1);
-
-    if (flags & kUIDropShadow) {
-        float offset = static_cast<float>(std::max(1, effectiveScale)) * scaleFactor;
-        color_t shadow = ColorA(color.a);
-        auto shadowPositions = makeQuad(fx + offset, fy + offset, fw, fh);
-        draw2d::submitQuad(shadowPositions, uvs, shadow.u32, texture);
-    }
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, uvs, color.u32, texture);
-    return x + w;
-}
-
-bool VulkanRenderer::getPicSize(int *w, int *h, qhandle_t pic) const {
-    auto it = images_.find(pic);
-    if (it == images_.end()) {
-        return false;
-    }
-
-    if (w) {
-        *w = it->second.width;
-    }
-    if (h) {
-        *h = it->second.height;
-    }
-    return it->second.transparent;
-}
-
-void VulkanRenderer::drawPic(int x, int y, color_t color, qhandle_t pic) {
-    auto it = images_.find(pic);
-    if (it == images_.end()) {
-        return;
-    }
-
-    if (it->second.width <= 0 || it->second.height <= 0) {
-        return;
-    }
-
-    drawStretchPic(x, y, it->second.width, it->second.height, color, pic);
-}
-
-void VulkanRenderer::drawStretchPic(int x, int y, int w, int h, color_t color, qhandle_t pic) {
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    if (!findImageRecord(pic)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, kFullUVs, color.u32, pic);
-}
-
-void VulkanRenderer::drawStretchRotatePic(int x, int y, int w, int h, color_t color, float angle, int pivot_x, int pivot_y, qhandle_t pic) {
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    if (!findImageRecord(pic)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    std::array<std::array<float, 2>, 4> positions = makeQuadPositions(static_cast<float>(x),
-                                                                       static_cast<float>(y),
-                                                                       static_cast<float>(x + w),
-                                                                       static_cast<float>(y + h));
-
-    float originX = static_cast<float>(x + pivot_x);
-    float originY = static_cast<float>(y + pivot_y);
-    float radians = DEG2RAD(angle);
-    float sine = std::sin(radians);
-    float cosine = std::cos(radians);
-
-    for (auto &pos : positions) {
-        float dx = pos[0] - originX;
-        float dy = pos[1] - originY;
-        float rotatedX = originX + dx * cosine - dy * sine;
-        float rotatedY = originY + dx * sine + dy * cosine;
-        pos[0] = rotatedX * scale;
-        pos[1] = rotatedY * scale;
-    }
-
-    draw2d::submitQuad(positions, kFullUVs, color.u32, pic);
-}
-
-void VulkanRenderer::drawKeepAspectPic(int x, int y, int w, int h, color_t color, qhandle_t pic) {
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    const ImageRecord *image = findImageRecord(pic);
-    if (!image) {
-        return;
-    }
-
-    if (image->flags & IF_SCRAP) {
-        drawStretchPic(x, y, w, h, color, pic);
-        return;
-    }
-
-    float imageWidth = static_cast<float>(std::max(1, image->width));
-    float imageHeight = static_cast<float>(std::max(1, image->height));
-    float aspect = imageWidth / imageHeight;
-
-    float destW = static_cast<float>(w);
-    float destH = static_cast<float>(h);
-    float scaledW = destW;
-    float scaledH = destH * aspect;
-    float scaleMax = std::max(scaledW, scaledH);
-    if (scaleMax <= 0.0f) {
-        return;
-    }
-
-    float s = 0.5f * (1.0f - (scaledW / scaleMax));
-    float t = 0.5f * (1.0f - (scaledH / scaleMax));
-
-    float scaleFactor = scale_;
-    if (scaleFactor <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scaleFactor;
-    float fy = static_cast<float>(y) * scaleFactor;
-    float fw = destW * scaleFactor;
-    float fh = destH * scaleFactor;
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    auto uvs = makeUV(s, t, 1.0f - s, 1.0f - t);
-    draw2d::submitQuad(positions, uvs, color.u32, pic);
-}
-
-void VulkanRenderer::drawStretchRaw(int x, int y, int w, int h) {
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    if (rawPic_.width <= 0 || rawPic_.height <= 0 || rawPic_.pixels.empty()) {
-        return;
-    }
-
-    qhandle_t texture = ensureRawTexture();
-    if (!texture || !findImageRecord(texture)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, kFullUVs, COLOR_WHITE.u32, texture);
-}
-
-void VulkanRenderer::updateRawPic(int pic_w, int pic_h, const uint32_t *pic) {
-    if (draw2d::isActive()) {
-        draw2d::flush();
-    }
-
-    if (pic_w <= 0 || pic_h <= 0 || !pic) {
-        rawPic_ = {};
-        if (auto it = images_.find(rawTextureHandle_); it != images_.end()) {
-            destroyImageRecord(it->second);
-        }
-        rawTextureHandle_ = 0;
-        return;
-    }
-
-    rawPic_.width = pic_w;
-    rawPic_.height = pic_h;
-    rawPic_.pixels.assign(pic, pic + (static_cast<size_t>(pic_w) * static_cast<size_t>(pic_h)));
-
-    if (rawTextureHandle_ != 0) {
-        if (auto it = images_.find(rawTextureHandle_); it != images_.end()) {
-            ImageRecord &record = it->second;
-            record.width = rawPic_.width;
-            record.height = rawPic_.height;
-            record.uploadWidth = rawPic_.width;
-            record.uploadHeight = rawPic_.height;
-            record.registrationSequence = r_registration_sequence;
-
-            const uint8_t *pixels = reinterpret_cast<const uint8_t *>(rawPic_.pixels.data());
-            size_t size = rawPic_.pixels.size() * sizeof(uint32_t);
-            uint32_t width = static_cast<uint32_t>(std::max(1, rawPic_.width));
-            uint32_t height = static_cast<uint32_t>(std::max(1, rawPic_.height));
-            ensureTextureResources(record, pixels, size, width, height, VK_FORMAT_R8G8B8A8_UNORM);
-            return;
-        }
-        rawTextureHandle_ = 0;
-    }
-
-    ensureRawTexture();
-}
-
-void VulkanRenderer::tileClear(int x, int y, int w, int h, qhandle_t pic) {
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    if (!findImageRecord(pic)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    float s0 = static_cast<float>(x) * kTileDivisor;
-    float t0 = static_cast<float>(y) * kTileDivisor;
-    float s1 = static_cast<float>(x + w) * kTileDivisor;
-    float t1 = static_cast<float>(y + h) * kTileDivisor;
-    auto uvs = makeUV(s0, t0, s1, t1);
-    draw2d::submitQuad(positions, uvs, COLOR_WHITE.u32, pic);
-}
-
-void VulkanRenderer::drawFill8(int x, int y, int w, int h, int c) {
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    qhandle_t texture = ensureWhiteTexture();
-    if (!texture || !findImageRecord(texture)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    color_t tint = ColorU32(d_8to24table[c & 0xFF]);
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, kFullUVs, tint.u32, texture);
-}
-
-void VulkanRenderer::drawFill32(int x, int y, int w, int h, color_t color) {
-    if (!frameActive_ || !draw2d::isActive()) {
-        return;
-    }
-  
-    if (!canSubmit2D()) {
-        return;
-    }
-
-    if (w <= 0 || h <= 0) {
-        return;
-    }
-
-    qhandle_t texture = ensureWhiteTexture();
-    if (!texture || !findImageRecord(texture)) {
-        return;
-    }
-
-    float scale = scale_;
-    if (scale <= 0.0f) {
-        return;
-    }
-
-    float fx = static_cast<float>(x) * scale;
-    float fy = static_cast<float>(y) * scale;
-    float fw = static_cast<float>(w) * scale;
-    float fh = static_cast<float>(h) * scale;
-
-    auto positions = makeQuad(fx, fy, fw, fh);
-    draw2d::submitQuad(positions, kFullUVs, color.u32, texture);
-}
-
-void VulkanRenderer::modeChanged(int width, int height, int flags) {
-    VideoGeometry geometry{};
-    geometry.width = std::max(1, width);
-    geometry.height = std::max(1, height);
-    geometry.flags = static_cast<vidFlags_t>(flags);
-
-    swapInterval_ = readSwapIntervalSetting();
-    if (swapInterval_ > 0) {
-        geometry.flags = static_cast<vidFlags_t>(geometry.flags | QVF_VIDEOSYNC);
-    } else {
-        geometry.flags = static_cast<vidFlags_t>(geometry.flags & ~QVF_VIDEOSYNC);
-    }
-
-    applyVideoGeometry(geometry);
-    updateUIScaling();
-
-    if (!initialized_ || device_ == VK_NULL_HANDLE || swapchain_ == VK_NULL_HANDLE) {
-        return;
-    }
-
-    if (geometry.width <= 0 || geometry.height <= 0) {
-        return;
-    }
-
-    rebuildSwapchain();
-}
-
-bool VulkanRenderer::videoSync() const {
-    if (!initialized_ || device_ == VK_NULL_HANDLE) {
-        return true;
-    }
-
-    if (!lastSubmittedFrame_.has_value() || inFlightFrames_.empty()) {
-        return true;
-    }
-
-    size_t index = *lastSubmittedFrame_;
-    if (index >= inFlightFrames_.size()) {
-        return true;
-    }
-
-    const InFlightFrame &frame = inFlightFrames_[index];
-    if (frame.inFlight == VK_NULL_HANDLE) {
-        return true;
-    }
-
-    VkResult status = vkGetFenceStatus(device_, frame.inFlight);
-    if (status == VK_SUCCESS) {
-        lastSubmittedFrame_.reset();
-        return true;
-    }
-
-    if (status == VK_NOT_READY) {
-        return false;
-    }
-
-    Com_Printf("refresh-vk: fence status check failed (VkResult %d).\n", static_cast<int>(status));
-    return true;
-}
-
-void VulkanRenderer::expireDebugObjects() {
-    R_ExpireDebugObjectsCPU();
-}
-
-bool VulkanRenderer::supportsPerPixelLighting() const {
-    return resolveToggle(vk_perPixelLighting, "gl_per_pixel_lighting", true) &&
-           resolveToggle(vk_dynamic, "gl_dynamic", true);
-}
-
-r_opengl_config_t VulkanRenderer::getGLConfig() const {
-    r_opengl_config_t cfg{};
-    cfg.colorbits = colorBitsForFormat(swapchainFormat_);
-    cfg.depthbits = 0;
-    cfg.stencilbits = 0;
-    cfg.multisamples = 0;
-    cfg.debug = supportsDebugUtils_ ? 1 : 0;
-    cfg.profile = QGL_PROFILE_NONE;
-    uint32_t apiVersion = physicalDeviceProperties_.apiVersion;
-    if (apiVersion == 0) {
-        apiVersion = VK_API_VERSION_1_0;
-    }
-    cfg.major_ver = static_cast<uint8_t>(VK_API_VERSION_MAJOR(apiVersion));
-    cfg.minor_ver = static_cast<uint8_t>(VK_API_VERSION_MINOR(apiVersion));
-    return cfg;
-}
-
-void VulkanRenderer::loadKFont(kfont_t *font, const char *filename) {
-    if (!font) {
-        return;
-    }
-
-    kfontCache_.erase(font);
-
-    RendererKFontLoadContext context{};
-    context.userData = this;
-    context.registerImage = [](void *userData, const char *path, imagetype_t type, imageflags_t flags) -> qhandle_t {
-        auto *renderer = static_cast<VulkanRenderer *>(userData);
-        return renderer->registerImage(path, type, flags);
-    };
-
-    RendererKFontData data{};
-    bool loaded = Renderer_LoadKFont(filename, context, &data);
-
-    if (!loaded) {
-        qhandle_t fallbackTexture = registerImage("_kfont", IT_FONT, IF_PERMANENT);
-        Renderer_BuildFallbackKFont(RENDERER_DEFAULT_KFONT_WIDTH, RENDERER_DEFAULT_KFONT_HEIGHT,
-                                    fallbackTexture, &data);
-    }
-
-    Renderer_AssignKFont(font, data);
-
-    if (loaded) {
-        KFontRecord record{};
-        record.texture = data.texture;
-        record.glyphs = data.glyphs;
-        record.lineHeight = data.lineHeight;
-        record.sw = data.sw;
-        record.sh = data.sh;
-        kfontCache_[font] = record;
-    }
-}
-
-const kfont_char_t *VulkanRenderer::lookupKFontChar(const kfont_t *kfont, uint32_t codepoint) const {
-    if (const KFontRecord *record = findKFontRecord(kfont)) {
-        return Renderer_LookupKFontGlyph(record->glyphs.data(), record->glyphs.size(), codepoint);
-    }
-
-    return Renderer_LookupKFontGlyph(kfont, codepoint);
-}
-
-void VulkanRenderer::resetFrameState() {
-    clearFrameTransientQueues();
-    resetFrameStatistics();
-
-    frameState_.refdef = {};
-    frameState_.entities.clear();
-    frameState_.dlights.clear();
-    frameState_.dynamicLights.clear();
-    frameState_.particles.clear();
-    frameState_.lightstyles.fill(lightstyle_t{});
-    frameState_.areaBits.clear();
-    frameState_.hasLightstyles = false;
-    frameState_.hasAreabits = false;
-    frameState_.hasRefdef = false;
-    frameState_.inWorldPass = false;
-    frameState_.worldRendered = false;
-    frameState_.dynamicLightsUploaded = false;
-    frameState_.skyActive = false;
-    frameState_.fogBits = FogNone;
-    frameState_.fogBitsSky = FogNone;
-    frameState_.fogEnabled = false;
-    frameState_.fogSkyEnabled = false;
-    frameState_.perPixelLighting = false;
-    frameState_.dynamicLightsEnabled = false;
-    frameState_.dynamicLightCount = 0;
-    frameState_.waterwarpActive = false;
-    frameState_.bloomActive = false;
-    frameState_.overlayBlendActive = false;
-}
-
-void VulkanRenderer::prepareFrameState(const refdef_t &fd) {
-    frameState_.refdef = fd;
-
-    frameState_.entities.clear();
-    frameState_.dynamicLights.clear();
-    frameState_.dynamicLightCount = 0;
-    frameState_.dynamicLightsUploaded = false;
-    if (fd.entities && fd.num_entities > 0) {
-        frameState_.entities.assign(fd.entities, fd.entities + fd.num_entities);
-        for (auto &ent : frameState_.entities) {
-            ent.next = nullptr;
-        }
-        frameState_.refdef.entities = frameState_.entities.data();
-        frameState_.refdef.num_entities = static_cast<int>(frameState_.entities.size());
-    } else {
-        frameState_.refdef.entities = nullptr;
-        frameState_.refdef.num_entities = 0;
-    }
-
-    frameState_.dlights.clear();
-    if (fd.dlights && fd.num_dlights > 0) {
-        frameState_.dlights.assign(fd.dlights, fd.dlights + fd.num_dlights);
-        frameState_.refdef.dlights = frameState_.dlights.data();
-        frameState_.refdef.num_dlights = static_cast<int>(frameState_.dlights.size());
-    } else {
-        frameState_.refdef.dlights = nullptr;
-        frameState_.refdef.num_dlights = 0;
-    }
-
-    frameState_.particles.clear();
-    if (fd.particles && fd.num_particles > 0) {
-        frameState_.particles.assign(fd.particles, fd.particles + fd.num_particles);
-        frameState_.refdef.particles = frameState_.particles.data();
-        frameState_.refdef.num_particles = static_cast<int>(frameState_.particles.size());
-    } else {
-        frameState_.refdef.particles = nullptr;
-        frameState_.refdef.num_particles = 0;
-    }
-
-    frameState_.hasLightstyles = false;
-    if (fd.lightstyles) {
-        std::copy(fd.lightstyles, fd.lightstyles + MAX_LIGHTSTYLES, frameState_.lightstyles.begin());
-        frameState_.refdef.lightstyles = frameState_.lightstyles.data();
-        frameState_.hasLightstyles = true;
-    } else {
-        frameState_.refdef.lightstyles = nullptr;
-    }
-
-    frameState_.areaBits.clear();
-    frameState_.hasAreabits = false;
-    if (fd.areabits) {
-        frameState_.areaBits.assign(fd.areabits, fd.areabits + MAX_MAP_AREA_BYTES);
-        frameState_.refdef.areabits = frameState_.areaBits.data();
-        frameState_.hasAreabits = true;
-    } else {
-        frameState_.refdef.areabits = nullptr;
-    }
-
-    frameState_.hasRefdef = true;
-}
-
-void VulkanRenderer::evaluateFrameSettings() {
-    frameState_.fogBits = FogNone;
-    frameState_.fogBitsSky = FogNone;
-
-    frameState_.fogEnabled = resolveToggle(vk_fog, "gl_fog", true);
-    if (frameState_.fogEnabled && frameState_.refdef.fog.density > 0.0f) {
-        frameState_.fogBits = static_cast<FogBits>(frameState_.fogBits | FogGlobal);
-    }
-
-    if (frameState_.fogEnabled && frameState_.refdef.heightfog.density > 0.0f && frameState_.refdef.heightfog.falloff > 0.0f) {
-        frameState_.fogBits = static_cast<FogBits>(frameState_.fogBits | FogHeight);
-    }
-
-    if (frameState_.fogEnabled && frameState_.refdef.fog.sky_factor > 0.0f) {
-        frameState_.fogBitsSky = static_cast<FogBits>(frameState_.fogBitsSky | FogSky);
-    }
-    frameState_.fogSkyEnabled = frameState_.fogBitsSky != FogNone;
-
-    bool dynamicAllowed = resolveToggle(vk_dynamic, "gl_dynamic", true);
-    bool perPixelAllowed = resolveToggle(vk_perPixelLighting, "gl_per_pixel_lighting", true);
-    frameState_.dynamicLightsEnabled = dynamicAllowed && perPixelAllowed;
-
-    if (!dynamicAllowed) {
-        frameState_.refdef.num_dlights = 0;
-    }
-
-    frameState_.perPixelLighting = frameState_.dynamicLightsEnabled;
-}
-
-void VulkanRenderer::uploadDynamicLights() {
-    frameState_.dynamicLights.clear();
-    frameState_.dynamicLightCount = 0;
-    frameState_.dynamicLightsUploaded = false;
-
-    if (!frameState_.dynamicLightsEnabled || frameState_.dlights.empty()) {
-        return;
-    }
-
-    size_t count = std::min(frameState_.dlights.size(), static_cast<size_t>(MAX_DLIGHTS));
-    frameState_.dynamicLights.reserve(count);
-
-    for (size_t i = 0; i < count; ++i) {
-        const dlight_t &src = frameState_.dlights[i];
-        FrameState::DynamicLightGPU gpu{};
-        gpu.positionRadius = { src.origin[0], src.origin[1], src.origin[2], src.radius };
-        gpu.colorIntensity = { src.color[0], src.color[1], src.color[2], src.intensity };
-        gpu.cone = { src.cone[0], src.cone[1], src.cone[2], src.cone[3] };
-        frameState_.dynamicLights.emplace_back(gpu);
-    }
-
-    frameState_.dynamicLightCount = static_cast<int>(count);
-    frameState_.dynamicLightsUploaded = count > 0;
-
-    if (frameState_.dynamicLightsUploaded) {
-        std::string label = "lighting.dynamic.upload";
-        label.push_back('(');
-        label.append(std::to_string(frameState_.dynamicLightCount));
-        label.push_back(')');
-        recordStage(label);
-    }
-}
-
-void VulkanRenderer::updateSkyState() {
-    frameState_.skyActive = !sky_.name.empty();
-}
-
-void VulkanRenderer::beginWorldPass() {
-    frameState_.inWorldPass = true;
-    frameState_.worldRendered = false;
-}
-
-void VulkanRenderer::renderWorld() {
-    frameState_.worldRendered = true;
-}
-
-void VulkanRenderer::endWorldPass() {
-    frameState_.inWorldPass = false;
-}
 
 } // namespace refresh::vk
